+++ coralnet_toolbox/Explorer/QtExplorer.py
@@ -1,43 +1,40 @@
+import os
 import warnings
 
-import os
-import gc
-
 import numpy as np
-
 import torch
-from ultralytics import YOLO
 
-from coralnet_toolbox.MachineLearning.Community.cfg import get_available_configs
-
-from coralnet_toolbox.Icons import get_icon
-from coralnet_toolbox.utilities import scale_pixmap
-from coralnet_toolbox.utilities import pixmap_to_numpy
-
-from PyQt5.QtGui import QIcon, QPen, QColor, QPainter, QImage, QBrush, QPainterPath, QPolygonF, QMouseEvent
-from PyQt5.QtCore import Qt, QTimer, QSize, QRect, QRectF, QPointF, pyqtSignal, QSignalBlocker, pyqtSlot
-
+from PyQt5.QtCore import Qt, QTimer, QRect, QRectF, QPointF, pyqtSignal, QSignalBlocker, pyqtSlot
+from PyQt5.QtGui import QPen, QColor, QPainter, QBrush, QPainterPath, QMouseEvent
 from PyQt5.QtWidgets import (QVBoxLayout, QHBoxLayout, QGraphicsView, QScrollArea,
                              QGraphicsScene, QPushButton, QComboBox, QLabel, QWidget, QGridLayout,
-                             QMainWindow, QSplitter, QGroupBox, QFormLayout,
-                             QSpinBox, QGraphicsEllipseItem, QGraphicsItem, QSlider,
-                             QListWidget, QDoubleSpinBox, QApplication, QStyle,
-                             QGraphicsRectItem, QRubberBand, QStyleOptionGraphicsItem,
-                             QTabWidget, QLineEdit, QFileDialog)
+                             QMainWindow, QSplitter, QGroupBox, QGraphicsItem, QSlider,
+                             QApplication, QGraphicsRectItem, QRubberBand)
 
+from ultralytics import YOLO
+
+from coralnet_toolbox.Icons import get_icon
 from coralnet_toolbox.QtProgressBar import ProgressBar
+from coralnet_toolbox.utilities import pixmap_to_numpy
+
+from .QtAnnotationDataItem import AnnotationDataItem
+from .QtAnnotationImageWidget import AnnotationImageWidget
+from .QtEmbeddingPointItem import EmbeddingPointItem
+from .QtSettingsWidgets import AnnotationSettingsWidget
+from .QtSettingsWidgets import EmbeddingSettingsWidget
+from .QtSettingsWidgets import ModelSettingsWidget
 
 try:
     from sklearn.preprocessing import StandardScaler
     from sklearn.decomposition import PCA
     from sklearn.manifold import TSNE
-    from umap import UMAP  
+    from umap import UMAP
 except ImportError:
     print("Warning: sklearn or umap not installed. Some features may be unavailable.")
     StandardScaler = None
     PCA = None
     TSNE = None
-    UMAP = None  
+    UMAP = None
 
 
 warnings.filterwarnings("ignore", category=DeprecationWarning)
@@ -47,305 +44,30 @@ warnings.filterwarnings("ignore", category=DeprecationWarning)
 # Constants
 # ----------------------------------------------------------------------------------------------------------------------
 
-ANNOTATION_WIDTH = 5
-
 POINT_SIZE = 15
 POINT_WIDTH = 3
 
-# ----------------------------------------------------------------------------------------------------------------------
-# Classes
-# ----------------------------------------------------------------------------------------------------------------------
-
-
-class AnnotationDataItem:
-    """Holds annotation information for consistent display across viewers."""
-
-    def __init__(self, annotation, embedding_x=None, embedding_y=None, embedding_id=None):
-        self.annotation = annotation
-        self.embedding_x = embedding_x if embedding_x is not None else 0.0
-        self.embedding_y = embedding_y if embedding_y is not None else 0.0
-        self.embedding_id = embedding_id if embedding_id is not None else 0
-        self._is_selected = False
-        self._preview_label = None
-        self._original_label = annotation.label
-        self._marked_for_deletion = False  # Track deletion status
-
-    @property
-    def effective_label(self):
-        """Get the current effective label (preview if exists, otherwise original)."""
-        return self._preview_label if self._preview_label else self.annotation.label
-
-    @property
-    def effective_color(self):
-        """Get the effective color for this annotation."""
-        return self.effective_label.color
-
-    @property
-    def is_selected(self):
-        """Check if this annotation is selected."""
-        return self._is_selected
-
-    def set_selected(self, selected):
-        """Set the selection state."""
-        self._is_selected = selected
-
-    def set_preview_label(self, label):
-        """Set a preview label for this annotation."""
-        self._preview_label = label
-
-    def clear_preview_label(self):
-        """Clear the preview label and revert to original."""
-        self._preview_label = None
-
-    def has_preview_changes(self):
-        """Check if this annotation has preview changes."""
-        return self._preview_label is not None
-
-    def mark_for_deletion(self):
-        """Mark this annotation for deletion."""
-        self._marked_for_deletion = True
-
-    def unmark_for_deletion(self):
-        """Unmark this annotation for deletion."""
-        self._marked_for_deletion = False
-
-    def is_marked_for_deletion(self):
-        """Check if this annotation is marked for deletion."""
-        return self._marked_for_deletion
-
-    def apply_preview_permanently(self):
-        """Apply the preview label permanently to the annotation."""
-        if self._preview_label:
-            self.annotation.update_label(self._preview_label)
-            self.annotation.update_user_confidence(self._preview_label)
-            self._original_label = self._preview_label
-            self._preview_label = None
-            return True
-        return False
-
-    def get_display_info(self):
-        """Get display information for this annotation."""
-        return {
-            'id': self.annotation.id,
-            'label': self.effective_label.short_label_code,
-            'confidence': self.get_effective_confidence(),
-            'type': type(self.annotation).__name__,
-            'image': os.path.basename(self.annotation.image_path),
-            'embedding_id': self.embedding_id,
-            'color': self.effective_color
-        }
-
-    def get_effective_confidence(self):
-        """Get the effective confidence value."""
-        if self.annotation.verified and hasattr(self.annotation, 'user_confidence') and self.annotation.user_confidence:
-            return list(self.annotation.user_confidence.values())[0]
-        elif hasattr(self.annotation, 'machine_confidence') and self.annotation.machine_confidence:
-            return list(self.annotation.machine_confidence.values())[0]
-        return 0.0
-    
-
-class AnnotationImageWidget(QWidget):
-    """Widget to display a single annotation image crop with selection support."""
-
-    def __init__(self, data_item, widget_height=96, annotation_viewer=None, parent=None):
-        super(AnnotationImageWidget, self).__init__(parent)
-        self.data_item = data_item
-        self.annotation = data_item.annotation
-        self.annotation_viewer = annotation_viewer
-        
-        self.widget_height = widget_height
-        self.aspect_ratio = 1.0  # Default to a square aspect ratio
-        self.pixmap = None       # Cache the original, unscaled pixmap
-
-        self.animation_offset = 0
-        self.animation_timer = QTimer(self)
-        self.animation_timer.timeout.connect(self._update_animation_frame)
-        self.animation_timer.setInterval(75)
-
-        self.setup_ui()
-        self.load_and_set_image() # Changed from load_annotation_image
-
-    def setup_ui(self):
-        """Set up the basic UI with a label for the image."""
-        layout = QVBoxLayout(self)
-        layout.setContentsMargins(4, 4, 4, 4)
-
-        self.image_label = QLabel()
-        self.image_label.setAlignment(Qt.AlignCenter)
-        self.image_label.setScaledContents(True)
-        self.image_label.setStyleSheet("border: none;")
-
-        layout.addWidget(self.image_label)
-
-    def load_and_set_image(self):
-        """Load image, calculate its aspect ratio, and set the widget's initial size."""
-        try:
-            # Use get_cropped_image_graphic() to get the QPixmap
-            cropped_pixmap = self.annotation.get_cropped_image_graphic()
-            if cropped_pixmap and not cropped_pixmap.isNull():
-                self.pixmap = cropped_pixmap
-                # Safely calculate aspect ratio
-                if self.pixmap.height() > 0:
-                    self.aspect_ratio = self.pixmap.width() / self.pixmap.height()
-                else:
-                    self.aspect_ratio = 1.0
-            else:
-                self.image_label.setText("No Image\nAvailable")
-                self.pixmap = None
-                self.aspect_ratio = 1.0
-        except Exception as e:
-            print(f"Error loading annotation image: {e}")
-            self.image_label.setText("Error\nLoading Image")
-            self.pixmap = None
-            self.aspect_ratio = 1.0
-        
-        # Trigger an initial size update
-        self.update_height(self.widget_height)
-
-    def update_height(self, new_height):
-        """Updates the widget's height and rescales its width and content accordingly."""
-        self.widget_height = new_height
-        
-        # Calculate the new width based on the stored aspect ratio
-        new_width = int(self.widget_height * self.aspect_ratio)
-
-        # Set the new fixed size for the entire widget
-        self.setFixedSize(new_width, new_height)
-
-        if self.pixmap:
-            # Scale the cached pixmap to fit the new widget size, leaving room for the border
-            # Note: We use the widget's new dimensions directly
-            scaled_pixmap = self.pixmap.scaled(
-                new_width - 8,  # Account for horizontal margins
-                new_height - 8, # Account for vertical margins
-                Qt.KeepAspectRatio,
-                Qt.SmoothTransformation
-            )
-            self.image_label.setPixmap(scaled_pixmap)
-        
-        self.update() # Schedule a repaint if needed
-
-    # Replace update_size with update_height
-    def update_size(self, new_size):
-        """Kept for compatibility, redirects to update_height."""
-        self.update_height(new_size)
-
-    def set_selected(self, selected):
-        """Set the selection state and update visual appearance."""
-        was_selected = self.is_selected()
-        
-        # Update the shared data item
-        self.data_item.set_selected(selected)
-
-        # Always update animation state, regardless of whether selection changed
-        if self.is_selected():
-            if not self.animation_timer.isActive():
-                self.animation_timer.start()
-        else:
-            if self.animation_timer.isActive():
-                self.animation_timer.stop()
-            self.animation_offset = 0
-
-        # Only trigger repaint if state actually changed or if we're selected
-        # (to ensure animation continues)
-        if was_selected != selected or selected:
-            self.update()
-
-    def is_selected(self):
-        """Return whether this widget is selected via the data item."""
-        return self.data_item.is_selected
-
-    def _update_animation_frame(self):
-        """Update the animation offset and schedule a repaint."""
-        self.animation_offset = (self.animation_offset + 1) % 20
-        self.update()
-
-    def paintEvent(self, event):
-        """Handle custom drawing for the widget, including the border."""
-        super().paintEvent(event)
-        painter = QPainter(self)
-        painter.setRenderHint(QPainter.Antialiasing)
-        
-        # Get the label that is currently active (which could be a preview)
-        effective_label = self.data_item.effective_label
-        
-        # Check if the active label is the special "Review" label (id == "-1")
-        if effective_label and effective_label.id == "-1":
-            # If it is, ALWAYS use a black pen for visibility against the white background.
-            pen_color = QColor("black")
-        else:
-            # Otherwise, use the effective color from the data item.
-            pen_color = self.data_item.effective_color
-
-        if self.is_selected():
-            pen = QPen(pen_color, ANNOTATION_WIDTH)
-            pen.setStyle(Qt.CustomDashLine)
-            pen.setDashPattern([2, 3])
-            pen.setDashOffset(self.animation_offset)
-        else:
-            pen = QPen(pen_color, ANNOTATION_WIDTH)  
-            pen.setStyle(Qt.SolidLine)
-        
-        painter.setPen(pen)
-        painter.setBrush(Qt.NoBrush)
-        
-        width = pen.width()
-        half_width = (width - 1) // 2
-        rect = self.rect().adjusted(half_width, half_width, -half_width, -half_width)
-        painter.drawRect(rect)
-
-    def mousePressEvent(self, event):
-        """Handle mouse press events for selection."""
-        if event.button() == Qt.LeftButton:
-            if self.annotation_viewer and hasattr(self.annotation_viewer, 'handle_annotation_selection'):
-                self.annotation_viewer.handle_annotation_selection(self, event)
-        # Ignore right mouse button clicks - don't call super() to prevent any selection
-        elif event.button() == Qt.RightButton:
-            event.ignore()
-            return
-        super().mousePressEvent(event)
-        
-
-class EmbeddingPointItem(QGraphicsEllipseItem):
-    """
-    A custom QGraphicsEllipseItem that prevents the default selection
-    rectangle from being drawn, and dynamically gets its color from the
-    shared AnnotationDataItem.
-    """
-    def paint(self, painter, option, widget):
-        # Get the shared data item, which holds the current state
-        data_item = self.data(0)
-        if data_item:
-            # Set the brush color based on the item's effective color
-            # This ensures preview colors are reflected instantly.
-            self.setBrush(data_item.effective_color)
-
-        # Remove the 'State_Selected' flag to prevent the default box
-        option.state &= ~QStyle.State_Selected
-        super(EmbeddingPointItem, self).paint(painter, option, widget)
-        
 
 # ----------------------------------------------------------------------------------------------------------------------
 # Viewers
 # ----------------------------------------------------------------------------------------------------------------------
 
-
 class EmbeddingViewer(QWidget):
     """Custom QGraphicsView for interactive embedding visualization with zooming, panning, and selection."""
-    
-    # Define signal to report selection changes
-    selection_changed = pyqtSignal(list)  # list of all currently selected annotation IDs
-    reset_view_requested = pyqtSignal()  # Signal to reset the view to fit all points
-    
+
+    # Define signals to report user actions to the ExplorerWindow controller
+    selection_changed = pyqtSignal(list)
+    reset_view_requested = pyqtSignal()
+
     def __init__(self, parent=None):
         # Create the graphics scene first
         self.graphics_scene = QGraphicsScene()
         self.graphics_scene.setSceneRect(-5000, -5000, 10000, 10000)
-        
+
         # Initialize as a QWidget
         super(EmbeddingViewer, self).__init__(parent)
         self.explorer_window = parent
-        
+
         # Create the actual graphics view
         self.graphics_view = QGraphicsView(self.graphics_scene)
         self.graphics_view.setRenderHint(QPainter.Antialiasing)
@@ -353,26 +75,26 @@ class EmbeddingViewer(QWidget):
         self.graphics_view.setHorizontalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
         self.graphics_view.setVerticalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
         self.graphics_view.setMinimumHeight(200)
-        
+
         # Custom rubber_band state variables
         self.rubber_band = None
         self.rubber_band_origin = QPointF()
         self.selection_at_press = None
-        
+
         self.points_by_id = {}  # Map annotation ID to embedding point
-        self.previous_selection_ids = set()  # Track previous selection to detect changes
-    
+        self.previous_selection_ids = set()
+
         self.animation_offset = 0
         self.animation_timer = QTimer()
         self.animation_timer.timeout.connect(self.animate_selection)
         self.animation_timer.setInterval(100)
-        
-        # Connect the scene's selection signal
-        self.graphics_scene.selectionChanged.connect(self.on_selection_changed)
-        
+
+        # Connect the scene's selection signal to our internal handler
+        self.graphics_scene.selectionChanged.connect(self._on_scene_selection_changed)
+
         # Setup the UI with header
         self.setup_ui()
-        
+
         # Connect mouse events to the graphics view
         self.graphics_view.mousePressEvent = self.mousePressEvent
         self.graphics_view.mouseDoubleClickEvent = self.mouseDoubleClickEvent
@@ -384,87 +106,58 @@ class EmbeddingViewer(QWidget):
         """Set up the UI with header layout and graphics view."""
         layout = QVBoxLayout(self)
         layout.setContentsMargins(0, 0, 0, 0)
-        
-        # Header layout
+
         header_layout = QHBoxLayout()
-        
-        # Home button
         self.home_button = QPushButton("Home")
         self.home_button.setToolTip("Reset view to fit all points")
-        self.home_button.clicked.connect(self.reset_view)
+        self.home_button.clicked.connect(self.fit_view_to_points)
         header_layout.addWidget(self.home_button)
-        
-        # Add stretch to push future controls to the right if needed
         header_layout.addStretch()
-        
+
         layout.addLayout(header_layout)
-        
-        # Add the graphics view
-        layout.addWidget(self.graphics_view)
-        # Add a placeholder label when no embedding is available
+
+        # Create a stacked layout for the view and placeholder
+        self.view_stack = QWidget()
+        stack_layout = QGridLayout(self.view_stack)
+        stack_layout.setContentsMargins(0, 0, 0, 0)
+
+        self.graphics_view.setParent(self.view_stack)
+        stack_layout.addWidget(self.graphics_view, 0, 0)
+
         self.placeholder_label = QLabel(
             "No embedding data available.\nPress 'Apply Embedding' to generate visualization."
         )
         self.placeholder_label.setAlignment(Qt.AlignCenter)
-        self.placeholder_label.setStyleSheet("color: gray; font-size: 14px;")
-        layout.addWidget(self.placeholder_label)
-        
-        # Initially show placeholder
+        self.placeholder_label.setStyleSheet("color: gray; font-size: 14px; background-color: #f0f0f0;")
+        self.placeholder_label.setParent(self.view_stack)
+        stack_layout.addWidget(self.placeholder_label, 0, 0)
+
+        layout.addWidget(self.view_stack)
         self.show_placeholder()
 
-    def reset_view(self):
-        """Reset the view to fit all embedding points."""
-        self.fit_view_to_points()
-
     def show_placeholder(self):
-        """Show the placeholder message and hide the graphics view."""
-        self.graphics_view.setVisible(False)
-        self.placeholder_label.setVisible(True)
+        self.graphics_view.hide()
+        self.placeholder_label.show()
         self.home_button.setEnabled(False)
 
     def show_embedding(self):
-        """Show the graphics view and hide the placeholder message."""
-        self.graphics_view.setVisible(True)
-        self.placeholder_label.setVisible(False)
+        self.placeholder_label.hide()
+        self.graphics_view.show()
         self.home_button.setEnabled(True)
-
-    # Delegate graphics view methods
-    def setRenderHint(self, hint):
-        self.graphics_view.setRenderHint(hint)
-    
-    def setDragMode(self, mode):
-        self.graphics_view.setDragMode(mode)
-    
-    def setTransformationAnchor(self, anchor):
-        self.graphics_view.setTransformationAnchor(anchor)
-    
-    def setResizeAnchor(self, anchor):
-        self.graphics_view.setResizeAnchor(anchor)
-    
-    def mapToScene(self, point):
-        return self.graphics_view.mapToScene(point)
-    
-    def scale(self, sx, sy):
-        self.graphics_view.scale(sx, sy)
-    
-    def translate(self, dx, dy):
-        self.graphics_view.translate(dx, dy)
-    
-    def fitInView(self, rect, aspect_ratio):
-        self.graphics_view.fitInView(rect, aspect_ratio)
 
     def mousePressEvent(self, event):
         """Handle mouse press for selection (point or rubber band) and panning."""
-        if event.button() == Qt.LeftButton and event.modifiers() == Qt.ControlModifier:
-            # Check if the click is on an existing point
+        if event.button() == Qt.LeftButton and \
+            (event.modifiers() == Qt.ControlModifier or
+             event.modifiers() == Qt.ShiftModifier):
+
             item_at_pos = self.graphics_view.itemAt(event.pos())
             if isinstance(item_at_pos, EmbeddingPointItem):
-                # If so, toggle its selection state and do nothing else
                 self.graphics_view.setDragMode(QGraphicsView.NoDrag)
-                item_at_pos.setSelected(not item_at_pos.isSelected())
-                return  # Event handled
+                # Let the default QGraphicsView logic handle Ctrl/Shift clicks on items
+                QGraphicsView.mousePressEvent(self.graphics_view, event)
+                return
 
-            # If the click was on the background, proceed with rubber band selection
             self.selection_at_press = set(self.graphics_scene.selectedItems())
             self.graphics_view.setDragMode(QGraphicsView.NoDrag)
             self.rubber_band_origin = self.graphics_view.mapToScene(event.pos())
@@ -474,92 +167,57 @@ class EmbeddingViewer(QWidget):
             self.graphics_scene.addItem(self.rubber_band)
 
         elif event.button() == Qt.RightButton:
-            # Handle panning
             self.graphics_view.setDragMode(QGraphicsView.ScrollHandDrag)
             left_event = QMouseEvent(event.type(), event.localPos(), Qt.LeftButton, Qt.LeftButton, event.modifiers())
             QGraphicsView.mousePressEvent(self.graphics_view, left_event)
         else:
-            # Handle standard single-item selection
             self.graphics_view.setDragMode(QGraphicsView.NoDrag)
             QGraphicsView.mousePressEvent(self.graphics_view, event)
-            
+
     def mouseDoubleClickEvent(self, event):
-        """Handle double-click to clear selection and reset the main view."""
         if event.button() == Qt.LeftButton:
-            # Clear selection if any items are selected
-            if self.graphics_scene.selectedItems():
-                self.graphics_scene.clearSelection()  # This triggers on_selection_changed
-            
-            # Signal the main window to revert from isolation mode
             self.reset_view_requested.emit()
             event.accept()
         else:
-            # Pass other double-clicks to the base class
-            super().mouseDoubleClickEvent(event)
+            QGraphicsView.mouseDoubleClickEvent(self.graphics_view, event)
 
     def mouseMoveEvent(self, event):
-        """Handle mouse move for dynamic selection and panning."""
         if self.rubber_band:
-            # Update the rubber band geometry
             current_pos = self.graphics_view.mapToScene(event.pos())
             self.rubber_band.setRect(QRectF(self.rubber_band_origin, current_pos).normalized())
-            
+
             path = QPainterPath()
             path.addRect(self.rubber_band.rect())
 
-            # Block signals to perform a compound selection operation
-            self.graphics_scene.blockSignals(True)
+            blocker = QSignalBlocker(self.graphics_scene)
+            # Use modifier to add to selection
+            self.graphics_scene.setSelectionArea(path, Qt.ControlModifier)
+            blocker.unblock()
 
-            # 1. Perform the "fancy" dynamic selection, which replaces the current selection
-            #    with only the items inside the rubber band.
-            self.graphics_scene.setSelectionArea(path)
-            
-            # 2. Add back the items that were selected at the start of the drag.
-            if self.selection_at_press:
-                for item in self.selection_at_press:
-                    item.setSelected(True)
-            
-            # Unblock signals and manually trigger our handler to process the final result.
-            self.graphics_scene.blockSignals(False)
-            self.on_selection_changed()
+            # Manually trigger our handler to process the final result.
+            self._on_scene_selection_changed()
 
         elif event.buttons() == Qt.RightButton:
-            # Handle right-click panning
-            left_event = QMouseEvent(event.type(), 
-                                     event.localPos(), 
-                                     Qt.LeftButton, 
-                                     Qt.LeftButton, 
-                                     event.modifiers())
+            left_event = QMouseEvent(event.type(), event.localPos(), Qt.LeftButton, Qt.LeftButton, event.modifiers())
             QGraphicsView.mouseMoveEvent(self.graphics_view, left_event)
         else:
             QGraphicsView.mouseMoveEvent(self.graphics_view, event)
 
     def mouseReleaseEvent(self, event):
-        """Handle mouse release to finalize the action and clean up."""
         if self.rubber_band:
-            # Clean up the visual rectangle
             self.graphics_scene.removeItem(self.rubber_band)
             self.rubber_band = None
-
-            # Clean up the stored selection state.
             self.selection_at_press = None
-            
+
         elif event.button() == Qt.RightButton:
-            # Finalize the pan
-            left_event = QMouseEvent(event.type(), 
-                                     event.localPos(), 
-                                     Qt.LeftButton, 
-                                     Qt.LeftButton, 
-                                     event.modifiers())
+            left_event = QMouseEvent(event.type(), event.localPos(), Qt.LeftButton, Qt.LeftButton, event.modifiers())
             QGraphicsView.mouseReleaseEvent(self.graphics_view, left_event)
             self.graphics_view.setDragMode(QGraphicsView.NoDrag)
         else:
-            # Finalize a single click
             QGraphicsView.mouseReleaseEvent(self.graphics_view, event)
             self.graphics_view.setDragMode(QGraphicsView.NoDrag)
-    
+
     def wheelEvent(self, event):
-        """Handle mouse wheel for zooming."""
         zoom_in_factor = 1.25
         zoom_out_factor = 1 / zoom_in_factor
 
@@ -570,224 +228,177 @@ class EmbeddingViewer(QWidget):
         zoom_factor = zoom_in_factor if event.angleDelta().y() > 0 else zoom_out_factor
         self.graphics_view.scale(zoom_factor, zoom_factor)
         new_pos = self.graphics_view.mapToScene(event.pos())
-        
+
         delta = new_pos - old_pos
         self.graphics_view.translate(delta.x(), delta.y())
 
     def update_embeddings(self, data_items):
-        """Update the embedding visualization with new data.
-
-        Args:
-            data_items: List of AnnotationDataItem objects.
-        """
         self.clear_points()
-        
+
         for item in data_items:
-            # Skip items marked for deletion
-            if item.is_marked_for_deletion():
-                continue
-                
             point = EmbeddingPointItem(0, 0, POINT_SIZE, POINT_SIZE)
             point.setPos(item.embedding_x, item.embedding_y)
-            
-            # No need to set initial brush - paint() will handle it
             point.setPen(QPen(QColor("black"), POINT_WIDTH))
-            
             point.setFlag(QGraphicsItem.ItemIgnoresTransformations)
             point.setFlag(QGraphicsItem.ItemIsSelectable)
-            
-            # This is the crucial link: store the shared AnnotationDataItem
             point.setData(0, item)
-            
+
             self.graphics_scene.addItem(point)
             self.points_by_id[item.annotation.id] = point
-            
-    def clear_points(self):
-        """Clear all embedding points from the scene."""
-        for point in self.points_by_id.values():
-            self.graphics_scene.removeItem(point)
-        self.points_by_id.clear()
 
-    def on_selection_changed(self):
-        """Handle point selection changes and emit a signal to the controller."""
-        # Check if graphics_scene still exists and is valid
-        if not self.graphics_scene or not hasattr(self.graphics_scene, 'selectedItems'):
-            return
-            
+    def clear_points(self):
+        self.graphics_scene.clear()
+        self.points_by_id.clear()
+        self.previous_selection_ids.clear()
+        # Add the rubberband back if it exists, though it should be None
+        if self.rubber_band:
+            self.graphics_scene.addItem(self.rubber_band)
+
+    @pyqtSlot()
+    def _on_scene_selection_changed(self):
+        """INTERNAL: Handles scene selection and emits a signal to the controller."""
         try:
             selected_items = self.graphics_scene.selectedItems()
-        except RuntimeError:
-            # Scene has been deleted
+            current_selection_ids = {item.data(0).annotation.id for item in selected_items}
+        except (RuntimeError, AttributeError):
             return
-            
-        current_selection_ids = {item.data(0).annotation.id for item in selected_items}
 
-        # If the selection has actually changed, update the model and emit
         if current_selection_ids != self.previous_selection_ids:
-            # Update the central model (the AnnotationDataItem) for all points
-            for point_id, point in self.points_by_id.items():
-                is_selected = point_id in current_selection_ids
-                point.data(0).set_selected(is_selected)
-
-            # Emit the complete list of currently selected IDs
-            self.selection_changed.emit(list(current_selection_ids))
             self.previous_selection_ids = current_selection_ids
+            self.selection_changed.emit(list(current_selection_ids))
+            self._update_animation_state()
 
-        # Handle local animation - check if animation_timer still exists
-        if hasattr(self, 'animation_timer') and self.animation_timer:
-            self.animation_timer.stop()
-            
+    def _update_animation_state(self):
+        """Starts or stops the selection animation timer."""
+        # Reset all pens first
         for point in self.points_by_id.values():
-            if not point.isSelected():
-                point.setPen(QPen(QColor("black"), POINT_WIDTH))
-        
-        if selected_items and hasattr(self, 'animation_timer') and self.animation_timer:
-            self.animation_timer.start()
+            point.setPen(QPen(QColor("black"), POINT_WIDTH))
+
+        if self.previous_selection_ids and self.points_by_id:
+            if not self.animation_timer.isActive():
+                self.animation_timer.start()
+        else:
+            if self.animation_timer.isActive():
+                self.animation_timer.stop()
 
     def animate_selection(self):
-        """Animate selected points with marching ants effect using darker versions of point colors."""
-        # Check if graphics_scene still exists and is valid
-        if not self.graphics_scene or not hasattr(self.graphics_scene, 'selectedItems'):
-            return
-            
-        try:
-            selected_items = self.graphics_scene.selectedItems()
-        except RuntimeError:
-            # Scene has been deleted
-            return
-            
         self.animation_offset = (self.animation_offset + 1) % 20
-        
-        # This logic remains the same. It applies the custom pen to the selected items.
-        # Because the items are EmbeddingPointItem, the default selection box won't be drawn.
-        for item in selected_items:
-            original_color = item.brush().color()
-            darker_color = original_color.darker(150)
+        for ann_id in self.previous_selection_ids:
+            if ann_id in self.points_by_id:
+                item = self.points_by_id[ann_id]
+                original_color = item.brush().color()
+                darker_color = original_color.darker(150)
 
-            animated_pen = QPen(darker_color, POINT_WIDTH)
-            animated_pen.setStyle(Qt.CustomDashLine)
-            animated_pen.setDashPattern([1, 2])
-            animated_pen.setDashOffset(self.animation_offset)
-            
-            item.setPen(animated_pen)
-            
+                animated_pen = QPen(darker_color, POINT_WIDTH)
+                animated_pen.setStyle(Qt.CustomDashLine)
+                animated_pen.setDashPattern([1, 2])
+                animated_pen.setDashOffset(self.animation_offset)
+                item.setPen(animated_pen)
+
+    def refresh_display(self, data_items, selected_ids):
+        """
+        PUBLIC: Re-renders the entire view based on the controller's data.
+        This is called after deletions or major state changes.
+        """
+        # Block signals to prevent feedback loops
+        blocker = QSignalBlocker(self.graphics_scene)
+
+        # 1. Update the points in the scene
+        self.update_embeddings(data_items)
+
+        # 2. Re-apply the selection
+        self.render_selection_from_ids(selected_ids)
+
+        # 3. Update view state
+        if not self.points_by_id:
+            self.show_placeholder()
+        else:
+            self.show_embedding()
+            # Optionally fit view if needed
+            self.fit_view_to_points()
+
     def render_selection_from_ids(self, selected_ids):
-        """Update the visual selection of points based on a set of IDs from the controller."""
-        # Filter out any IDs that are marked for deletion
-        visible_selected_ids = set()
+        """PUBLIC: Updates the visual selection of points from a set of IDs."""
+        blocker = QSignalBlocker(self.graphics_scene)
+
+        self.graphics_scene.clearSelection()
         for ann_id in selected_ids:
             if ann_id in self.points_by_id:
-                point = self.points_by_id[ann_id]
-                data_item = point.data(0)
-                if data_item and not data_item.is_marked_for_deletion():
-                    visible_selected_ids.add(ann_id)
-        
-        # Block this scene's own selectionChanged signal to prevent an infinite loop
-        blocker = QSignalBlocker(self.graphics_scene)
-        
-        for ann_id, point in self.points_by_id.items():
-            data_item = point.data(0)
-            # Only show selection for items not marked for deletion
-            if data_item and not data_item.is_marked_for_deletion():
-                point.setSelected(ann_id in visible_selected_ids)
-            else:
-                point.setSelected(False)
-            
-        self.previous_selection_ids = visible_selected_ids
-        
-        # Trigger animation update
-        self.on_selection_changed()
+                self.points_by_id[ann_id].setSelected(True)
+
+        self.previous_selection_ids = set(selected_ids)
+        self._update_animation_state()
+
+    def refresh_point_colors(self):
+        """PUBLIC: Forces all points to repaint, updating their colors from data_items."""
+        for point in self.points_by_id.values():
+            point.update()
 
     def fit_view_to_points(self):
-        """Fit the view to show all embedding points."""
         if self.points_by_id:
             self.graphics_view.fitInView(self.graphics_scene.itemsBoundingRect(), Qt.KeepAspectRatio)
-        else:
-            # If no points, reset to default view
-            self.graphics_view.fitInView(-2500, -2500, 5000, 5000, Qt.KeepAspectRatio)
-            
 
-class AnnotationViewer(QScrollArea):
-    """Scrollable grid widget for displaying annotation image crops with selection,
-    filtering, and isolation support.
+
+class AnnotationViewer(QWidget):
     """
-    
-    # Define signals to report changes to the ExplorerWindow
-    selection_changed = pyqtSignal(list)  # list of changed annotation IDs
-    preview_changed = pyqtSignal(list)   # list of annotation IDs with new previews
-    reset_view_requested = pyqtSignal()  # Signal to reset the view to fit all points
+    A compound widget with a toolbar and a scrollable grid for displaying annotation crops.
+    It inherits from QWidget to properly manage its child widgets.
+    """
+
+    # Signals to report user actions to the ExplorerWindow controller
+    selection_changed = pyqtSignal(list)
+    deletion_requested = pyqtSignal(list)
+    preview_label_applied = pyqtSignal()
+    reset_view_requested = pyqtSignal()
 
     def __init__(self, parent=None):
         super(AnnotationViewer, self).__init__(parent)
         self.annotation_widgets_by_id = {}
-        self.selected_widgets = []
-        self.last_selected_index = -1
+        self.current_selection_ids = set()
+        self.last_selected_id = None
         self.current_widget_size = 96
-        
-        self.selection_at_press = set()
+
         self.rubber_band = None
         self.rubber_band_origin = None
-        self.drag_threshold = 5
-        self.mouse_pressed_on_widget = False
-        
-        self.preview_label_assignments = {}
-        self.original_label_assignments = {}
-        
-        # New state variables for Isolate/Focus mode
+
         self.isolated_mode = False
-        self.isolated_widgets = set()
+        self.isolated_ids = set()
 
         self.setup_ui()
 
     def setup_ui(self):
         """Set up the UI with a toolbar and a scrollable content area."""
-        self.setWidgetResizable(True)
-        self.setHorizontalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
-        self.setVerticalScrollBarPolicy(Qt.ScrollBarAsNeeded)
-
-        # Main container and layout
-        main_container = QWidget()
-        main_layout = QVBoxLayout(main_container)
+        main_layout = QVBoxLayout(self)
         main_layout.setContentsMargins(0, 0, 0, 0)
-        main_layout.setSpacing(4)  # Add a little space between toolbar and content
+        main_layout.setSpacing(4)
 
-        # --- New Toolbar ---
+        # --- Toolbar ---
         toolbar_widget = QWidget()
         toolbar_layout = QHBoxLayout(toolbar_widget)
         toolbar_layout.setContentsMargins(4, 2, 4, 2)
 
-        # Isolate/Focus controls
         self.isolate_button = QPushButton("Isolate Selection")
-        isolate_icon = get_icon("focus.png")
-        if not isolate_icon.isNull():
-            self.isolate_button.setIcon(isolate_icon)
+        self.isolate_button.setIcon(get_icon("focus.png"))
         self.isolate_button.setToolTip("Hide all non-selected annotations")
         self.isolate_button.clicked.connect(self.isolate_selection)
         toolbar_layout.addWidget(self.isolate_button)
 
         self.show_all_button = QPushButton("Show All")
-        show_all_icon = get_icon("show_all.png")
-        if not show_all_icon.isNull():
-            self.show_all_button.setIcon(show_all_icon)
+        self.show_all_button.setIcon(get_icon("show_all.png"))
         self.show_all_button.setToolTip("Show all filtered annotations")
         self.show_all_button.clicked.connect(self.show_all_annotations)
         toolbar_layout.addWidget(self.show_all_button)
 
-        # Add a separator
         toolbar_layout.addWidget(self._create_separator())
 
-        # Sort controls
         sort_label = QLabel("Sort By:")
         toolbar_layout.addWidget(sort_label)
         self.sort_combo = QComboBox()
         self.sort_combo.addItems(["None", "Label", "Image"])
-        self.sort_combo.currentTextChanged.connect(self.on_sort_changed)
+        self.sort_combo.currentTextChanged.connect(self.recalculate_widget_positions)
         toolbar_layout.addWidget(self.sort_combo)
-
-        # Add a spacer to push the size controls to the right
         toolbar_layout.addStretch()
 
-        # Size controls
         size_label = QLabel("Size:")
         toolbar_layout.addWidget(size_label)
         self.size_slider = QSlider(Qt.Horizontal)
@@ -804,1263 +415,373 @@ class AnnotationViewer(QScrollArea):
         toolbar_layout.addWidget(self.size_value_label)
 
         main_layout.addWidget(toolbar_widget)
-        
-        # --- Content Area ---
-        self.content_widget = QWidget()
-        content_scroll = QScrollArea()
-        content_scroll.setWidgetResizable(True)
-        content_scroll.setHorizontalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
-        content_scroll.setVerticalScrollBarPolicy(Qt.ScrollBarAsNeeded)
-        content_scroll.setWidget(self.content_widget)
-        
-        main_layout.addWidget(content_scroll)
-        self.setWidget(main_container)
 
-        # Set the initial state of the toolbar buttons
+        # --- Scroll Area ---
+        self.scroll_area = QScrollArea()
+        self.scroll_area.setWidgetResizable(True)
+        self.scroll_area.setHorizontalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
+        self.scroll_area.setVerticalScrollBarPolicy(Qt.ScrollBarAsNeeded)
+
+        self.content_widget = QWidget()
+        self.scroll_area.setWidget(self.content_widget)
+
+        # Install an event filter on the scroll area's viewport to capture mouse events
+        self.scroll_area.viewport().installEventFilter(self)
+
+        main_layout.addWidget(self.scroll_area)
         self._update_toolbar_state()
-            
+
+    def eventFilter(self, source, event):
+        """
+        Filters events from the scroll area's viewport to handle mouse-based
+        selection and rubber banding on the background.
+        """
+        if source is self.scroll_area.viewport():
+            if event.type() == QMouseEvent.MouseButtonPress:
+                self.mousePressEvent(event)
+                return True
+            elif event.type() == QMouseEvent.MouseButtonDblClick:
+                self.mouseDoubleClickEvent(event)
+                return True
+            elif event.type() == QMouseEvent.MouseMove:
+                self.mouseMoveEvent(event)
+                return True
+            elif event.type() == QMouseEvent.MouseButtonRelease:
+                self.mouseReleaseEvent(event)
+                return True
+
+        return super(AnnotationViewer, self).eventFilter(source, event)
+
     @pyqtSlot()
     def isolate_selection(self):
-        """Hides all annotation widgets that are not currently selected."""
-        if not self.selected_widgets or self.isolated_mode:
+        if not self.current_selection_ids or self.isolated_mode:
             return
-
-        self.isolated_widgets = set(self.selected_widgets)
-        self.content_widget.setUpdatesEnabled(False)
-        try:
-            for widget in self.annotation_widgets_by_id.values():
-                if widget not in self.isolated_widgets:
-                    widget.hide()
-            self.isolated_mode = True
-            self.recalculate_widget_positions()
-        finally:
-            self.content_widget.setUpdatesEnabled(True)
-
+        self.isolated_mode = True
+        self.isolated_ids = self.current_selection_ids.copy()
+        self.recalculate_widget_positions()
         self._update_toolbar_state()
 
     @pyqtSlot()
     def show_all_annotations(self):
-        """Shows all annotation widgets, exiting the isolated mode."""
         if not self.isolated_mode:
             return
-            
         self.isolated_mode = False
-        self.isolated_widgets.clear()
-        
-        self.content_widget.setUpdatesEnabled(False)
-        try:
-            for widget in self.annotation_widgets_by_id.values():
-                # Only show widgets that are not marked for deletion
-                if not widget.data_item.is_marked_for_deletion():
-                    widget.show()
-            self.recalculate_widget_positions()
-        finally:
-            self.content_widget.setUpdatesEnabled(True)
-            
+        self.isolated_ids.clear()
+        self.recalculate_widget_positions()
         self._update_toolbar_state()
-        
+
     def _update_toolbar_state(self):
-        """Updates the visibility and enabled state of the toolbar buttons
-        based on the current selection and isolation mode.
-        """
-        selection_exists = bool(self.selected_widgets)
-        
+        selection_exists = bool(self.current_selection_ids)
         if self.isolated_mode:
             self.isolate_button.hide()
             self.show_all_button.show()
-            self.show_all_button.setEnabled(True)
         else:
             self.isolate_button.show()
             self.show_all_button.hide()
             self.isolate_button.setEnabled(selection_exists)
-            
+
     def _create_separator(self):
-        """Create a vertical separator line for the toolbar."""
         separator = QLabel("|")
         separator.setStyleSheet("color: gray; margin: 0 5px;")
         return separator
 
-    def on_sort_changed(self, sort_type):
-        """Handle sort type change."""
-        self.recalculate_widget_positions()
-
     def _get_sorted_widgets(self):
-        """Get widgets sorted according to the current sort setting, excluding deleted ones."""
+        """Gets a sorted list of currently visible widgets."""
         sort_type = self.sort_combo.currentText()
-        
-        # Only include widgets not marked for deletion
-        widgets = [w for w in self.annotation_widgets_by_id.values() 
-                  if not w.data_item.is_marked_for_deletion()]
-        
+
+        widgets = [w for w in self.annotation_widgets_by_id.values()if not w.isHidden()]
+
         if sort_type == "None":
             return widgets
-        
+
+        key_func = None
         if sort_type == "Label":
-            widgets.sort(key=lambda w: w.data_item.effective_label.short_label_code)
+            def key_func(w):
+                return w.data_item.effective_label.short_label_code
         elif sort_type == "Image":
-            widgets.sort(key=lambda w: os.path.basename(w.data_item.annotation.image_path))
-        
+            def key_func(w): return os.path.basename(
+                w.data_item.annotation.image_path)
+
+        if key_func:
+            widgets.sort(key=key_func)
+
         return widgets
 
     def _group_widgets_by_sort_key(self, widgets):
-        """Group widgets by the current sort key and return groups with headers."""
         sort_type = self.sort_combo.currentText()
-        
         if sort_type == "None":
             return [("", widgets)]
-        
+
         groups = []
-        current_group = []
-        current_key = None
-        
+        current_group, current_key = [], None
         for widget in widgets:
+            key = ""
             if sort_type == "Label":
                 key = widget.data_item.effective_label.short_label_code
             elif sort_type == "Image":
                 key = os.path.basename(widget.data_item.annotation.image_path)
-            else:
-                key = ""
-            
+
             if current_key != key:
                 if current_group:
                     groups.append((current_key, current_group))
-                current_group = [widget]
-                current_key = key
+                current_group, current_key = [widget], key
             else:
                 current_group.append(widget)
-        
         if current_group:
             groups.append((current_key, current_group))
-        
         return groups
-    
+
     def _clear_separator_labels(self):
-        """Remove any existing group header labels."""
         if hasattr(self, '_group_headers'):
             for header in self._group_headers:
-                header.setParent(None)
                 header.deleteLater()
         self._group_headers = []
 
     def _create_group_header(self, text):
-        """Create a group header label."""
         if not hasattr(self, '_group_headers'):
             self._group_headers = []
-        
-        header = QLabel(text)
-        header.setParent(self.content_widget)
-        header.setStyleSheet("""
-            QLabel {
-                font-weight: bold;
-                font-size: 12px;
-                color: #555;
-                background-color: #f0f0f0;
-                border: 1px solid #ccc;
-                border-radius: 3px;
-                padding: 5px 8px;
-                margin: 2px 0px;
-            }
-        """)
-        header.setFixedHeight(30)  # Increased from 25 to 30
-        header.setMinimumWidth(self.viewport().width() - 20)
-        header.show()
-        
+        header = QLabel(text, self.content_widget)
+        header.setStyleSheet(
+            "font-weight: bold; "
+            "font-size: 12px; "
+            "color: #555; "
+            "background-color: #f0f0f0; "
+            "border: 1px solid #ccc; "
+            "border-radius: 3px; "
+            "padding: 5px 8px; "
+            "margin: 2px 0px;")
+        header.setFixedHeight(30)
+        header.setMinimumWidth(self.scroll_area.viewport().width() - 20)
         self._group_headers.append(header)
         return header
 
     def on_size_changed(self, value):
-        """Handle slider value change to resize annotation widgets."""
         if value % 2 != 0:
             value -= 1
         self.current_widget_size = value
         self.size_value_label.setText(str(value))
-        
-        # Disable updates for performance while resizing many items
+
         self.content_widget.setUpdatesEnabled(False)
         for widget in self.annotation_widgets_by_id.values():
-            widget.update_height(value)  # Call the new, more descriptive method
+            widget.update_height(value)
         self.content_widget.setUpdatesEnabled(True)
-
-        # After resizing, reflow the layout
         self.recalculate_widget_positions()
 
-    def recalculate_grid_layout(self):
-        """Recalculate the grid layout based on current widget width."""
-        if not self.annotation_widgets_by_id:
-            return
-            
-        available_width = self.viewport().width() - 20
-        widget_width = self.current_widget_size + self.grid_layout.spacing()
-        cols = max(1, available_width // widget_width)
-        
-        for i, widget in enumerate(self.annotation_widgets_by_id.values()):
-            self.grid_layout.addWidget(widget, i // cols, i % cols)
-            
     def recalculate_widget_positions(self):
         """Manually positions widgets in a flow layout with sorting and group headers."""
-        if not self.annotation_widgets_by_id:
-            self.content_widget.setMinimumSize(1, 1)
-            return
-
-        # Clear any existing separator labels
+        self.content_widget.setUpdatesEnabled(False)
         self._clear_separator_labels()
 
-        # Get sorted widgets that are not marked for deletion
-        all_widgets = self._get_sorted_widgets()
-        
-        # Filter to only visible widgets (not hidden and not marked for deletion)
-        visible_widgets = [w for w in all_widgets if not w.isHidden() and not w.data_item.is_marked_for_deletion()]
-        
-        if not visible_widgets:
-            self.content_widget.setMinimumSize(1, 1)
-            return
+        try:
+            for ann_id, widget in self.annotation_widgets_by_id.items():
+                is_deleted = widget.data_item.is_marked_for_deletion()
+                in_isolated_view = self.isolated_mode and ann_id in self.isolated_ids
+                should_be_visible = not is_deleted and (not self.isolated_mode or in_isolated_view)
+                widget.setVisible(should_be_visible)
 
-        # Group widgets by sort key
-        groups = self._group_widgets_by_sort_key(visible_widgets)
+            visible_widgets = self._get_sorted_widgets()
+            if not visible_widgets:
+                self.content_widget.setMinimumSize(1, 1)
+                return
 
-        # Calculate spacing
-        spacing = max(5, int(self.current_widget_size * 0.08))
-        available_width = self.viewport().width()
-        
-        x, y = spacing, spacing
-        max_height_in_row = 0
+            groups = self._group_widgets_by_sort_key(visible_widgets)
+            spacing = max(5, int(self.current_widget_size * 0.08))
+            available_width = self.scroll_area.viewport().width()
+            x, y = spacing, spacing
+            max_height_in_row = 0
 
-        for group_name, group_widgets in groups:
-            # Add group header if sorting is enabled and group has a name
-            if group_name and self.sort_combo.currentText() != "None":
-                # Ensure we're at the start of a new line for headers
-                if x > spacing:
+            for group_name, group_widgets in groups:
+                if group_name and self.sort_combo.currentText() != "None":
+                    if x > spacing:
+                        x, y = spacing, y + max_height_in_row + spacing
+                        max_height_in_row = 0
+                    header_label = self._create_group_header(group_name)
+                    header_label.move(x, y)
+                    header_label.show()
+                    y += header_label.height() + spacing
                     x = spacing
-                    y += max_height_in_row + spacing
-                    max_height_in_row = 0
-                
-                # Create and position header label
-                header_label = self._create_group_header(group_name)
-                header_label.move(x, y)
-                
-                # Move to next line after header
-                y += header_label.height() + spacing
-                x = spacing
-                max_height_in_row = 0
 
-            # Position widgets in this group
-            for widget in group_widgets:
-                widget_size = widget.size()
-                
-                # Check if widget fits on current line
-                if x > spacing and x + widget_size.width() > available_width:
-                    x = spacing
-                    y += max_height_in_row + spacing
-                    max_height_in_row = 0
+                for widget in group_widgets:
+                    widget_size = widget.size()
+                    if x > spacing and x + widget_size.width() > available_width:
+                        x, y = spacing, y + max_height_in_row + spacing
+                        max_height_in_row = 0
+                    widget.move(x, y)
+                    x += widget_size.width() + spacing
+                    max_height_in_row = max(
+                        max_height_in_row, widget_size.height())
 
-                widget.move(x, y)
-                x += widget_size.width() + spacing
-                max_height_in_row = max(max_height_in_row, widget_size.height())
+            total_height = y + max_height_in_row + spacing
+            self.content_widget.setMinimumSize(available_width, total_height)
+        finally:
+            self.content_widget.setUpdatesEnabled(True)
+            self.content_widget.update()
 
-        # Update content widget size
-        total_height = y + max_height_in_row + spacing
-        self.content_widget.setMinimumSize(available_width, total_height)
-            
     def update_annotations(self, data_items):
-        """Update displayed annotations, creating new widgets. This will also
-        reset any active isolation view.
-        """
-        # Reset isolation state before updating to avoid confusion
-        if self.isolated_mode:
-            self.show_all_annotations()
-            
-        # Clear any existing widgets and ensure they are deleted
+        """PUBLIC: Rebuilds the view from a fresh list of data items."""
+        self.isolated_mode = False
+        self.isolated_ids.clear()
+
         for widget in self.annotation_widgets_by_id.values():
             widget.setParent(None)
             widget.deleteLater()
-            
-        self.annotation_widgets_by_id.clear()
-        self.selected_widgets.clear()
-        self.last_selected_index = -1
 
-        # Create new widgets, parenting them to the content_widget
+        self.annotation_widgets_by_id.clear()
+        self.current_selection_ids.clear()
+        self.last_selected_id = None
+
         for data_item in data_items:
-            annotation_widget = AnnotationImageWidget(
-                data_item, 
-                self.current_widget_size,
-                annotation_viewer=self,
-                parent=self.content_widget
-            )
-            annotation_widget.show() 
-            self.annotation_widgets_by_id[data_item.annotation.id] = annotation_widget
-        
+            widget = AnnotationImageWidget(data_item, self.current_widget_size, self, self.content_widget)
+            self.annotation_widgets_by_id[data_item.annotation.id] = widget
+
         self.recalculate_widget_positions()
-        # Ensure toolbar is in the correct state after a refresh
         self._update_toolbar_state()
 
     def resizeEvent(self, event):
-        """On window resize, reflow the annotation widgets."""
         super().resizeEvent(event)
-        # Use a QTimer to avoid rapid, expensive reflows while dragging the resize handle
         if not hasattr(self, '_resize_timer'):
             self._resize_timer = QTimer(self)
             self._resize_timer.setSingleShot(True)
             self._resize_timer.timeout.connect(self.recalculate_widget_positions)
-        # Restart the timer on each resize event
-        self._resize_timer.start(100)  # 100ms delay
-        
+        self._resize_timer.start(50)
+
     def mousePressEvent(self, event):
-        """Handle mouse press for starting rubber band selection OR clearing selection."""
-        
-        # Handle plain left-clicks
+        # This event is now called from the event filter
         if event.button() == Qt.LeftButton:
-            
-            # This is the new logic for clearing selection on a background click.
-            if not event.modifiers():  # Check for NO modifiers (e.g., Ctrl, Shift)
-                
-                is_on_widget = False
-                child_at_pos = self.childAt(event.pos())
-
-                # Determine if the click was on an actual annotation widget or empty space
-                if child_at_pos:
-                    widget = child_at_pos
-                    while widget and widget != self:
-                        if hasattr(widget, 'annotation_viewer') and widget.annotation_viewer == self:
-                            is_on_widget = True
-                            break
-                        widget = widget.parent()
-                
-                # If click was on empty space AND something is currently selected...
-                if not is_on_widget and self.selected_widgets:
-                    # Get IDs of widgets that are about to be deselected to emit a signal
-                    changed_ids = [w.data_item.annotation.id for w in self.selected_widgets]
-                    self.clear_selection()
-                    self.selection_changed.emit(changed_ids)
-                    # The event is handled, but we don't call super() to prevent
-                    # the scroll area from doing anything else, like starting a drag.
-                    return
-
-            # Handle Ctrl+Click for rubber band
-            elif event.modifiers() == Qt.ControlModifier:
-                # Store the set of currently selected items.
-                self.selection_at_press = set(self.selected_widgets)
-                self.rubber_band_origin = event.pos()
-                # We determine mouse_pressed_on_widget here but use it in mouseMove
-                self.mouse_pressed_on_widget = False
-                child_widget = self.childAt(event.pos())
-                if child_widget:
-                    widget = child_widget
-                    while widget and widget != self:
-                        if hasattr(widget, 'annotation_viewer') and widget.annotation_viewer == self:
-                            self.mouse_pressed_on_widget = True
-                            break
-                        widget = widget.parent()
+            # Check if click was on the background of the viewport
+            if self.content_widget.childAt(event.pos()) is None:
+                if event.modifiers() == Qt.NoModifier and self.current_selection_ids:
+                    # Signal to clear selection
+                    self.selection_changed.emit([])
+                elif event.modifiers() == Qt.ControlModifier:
+                    self.rubber_band_origin = event.pos()
+                    self.rubber_band = QRubberBand(QRubberBand.Rectangle, self.scroll_area.viewport())
+                    self.rubber_band.move(self.rubber_band_origin)
+                    self.rubber_band.show()
+                event.accept()
                 return
-                
-        # Handle right-clicks
-        elif event.button() == Qt.RightButton:
-            event.ignore()
-            return
-            
-        # For all other cases (e.g., a click on a widget that should be handled
-        # by the widget itself), pass the event to the default handler.
         super().mousePressEvent(event)
-        
-    def mouseDoubleClickEvent(self, event):
-        """Handle double-click to clear selection and exit isolation mode."""
-        if event.button() == Qt.LeftButton:
-            changed_ids = []
-            
-            # If items are selected, clear the selection and record their IDs
-            if self.selected_widgets:
-                changed_ids = [w.data_item.annotation.id for w in self.selected_widgets]
-                self.clear_selection()
-                self.selection_changed.emit(changed_ids)
 
-            # If in isolation mode, revert to showing all annotations
-            if self.isolated_mode:
-                self.show_all_annotations()
-            
-            # Signal the main window to reset its view (e.g., switch tabs)
+    def mouseDoubleClickEvent(self, event):
+        # This event is now called from the event filter
+        if event.button() == Qt.LeftButton and self.content_widget.childAt(event.pos()) is None:
             self.reset_view_requested.emit()
             event.accept()
         else:
             super().mouseDoubleClickEvent(event)
-    
+
     def mouseMoveEvent(self, event):
-        """Handle mouse move for DYNAMIC rubber band selection."""
-        if self.rubber_band_origin is None or \
-           event.buttons() != Qt.LeftButton or \
-           event.modifiers() != Qt.ControlModifier:
-            super().mouseMoveEvent(event)
+        # This event is now called from the event filter
+        if self.rubber_band and self.rubber_band.isVisible():
+            self.rubber_band.setGeometry(
+                QRect(self.rubber_band_origin, event.pos()).normalized())
+
+            selection_rect = self.rubber_band.geometry()
+            newly_selected_ids = set()
+
+            for widget in self.annotation_widgets_by_id.values():
+                if widget.isVisible() and selection_rect.intersects(widget.geometry()):
+                    newly_selected_ids.add(widget.data_item.annotation.id)
+
+            # Combine with Ctrl-held items
+            if QApplication.keyboardModifiers() & Qt.ControlModifier:
+                # This part is complex; for now, we just update based on rubber band
+                pass
+
+            self.selection_changed.emit(list(newly_selected_ids))
             return
+        super().mouseMoveEvent(event)
 
-        # If the mouse was pressed on a widget, let that widget handle the event.
-        if self.mouse_pressed_on_widget:
-            super().mouseMoveEvent(event)
-            return
-
-        # Only start the rubber band after dragging a minimum distance
-        distance = (event.pos() - self.rubber_band_origin).manhattanLength()
-        if distance < self.drag_threshold:
-            return
-
-        # Create and show the rubber band if it doesn't exist
-        if not self.rubber_band:
-            self.rubber_band = QRubberBand(QRubberBand.Rectangle, self.viewport())
-        
-        rect = QRect(self.rubber_band_origin, event.pos()).normalized()
-        self.rubber_band.setGeometry(rect)
-        self.rubber_band.show()
-
-        # Perform dynamic selection on every move
-        selection_rect = self.rubber_band.geometry()
-        content_widget = self.content_widget
-        changed_ids = []
-
-        for widget in self.annotation_widgets_by_id.values():
-            widget_rect_in_content = widget.geometry()
-            # Map widget's geometry from the content area to the visible viewport
-            widget_rect_in_viewport = QRect(
-                content_widget.mapTo(self.viewport(), widget_rect_in_content.topLeft()),
-                widget_rect_in_content.size()
-            )
-
-            is_in_band = selection_rect.intersects(widget_rect_in_viewport)
-            
-            # A widget should be selected if it was selected at the start OR is in the band now.
-            should_be_selected = (widget in self.selection_at_press) or is_in_band
-
-            if should_be_selected and not widget.is_selected():
-                if self.select_widget(widget):
-                    changed_ids.append(widget.data_item.annotation.id)
-            elif not should_be_selected and widget.is_selected():
-                if self.deselect_widget(widget):
-                    changed_ids.append(widget.data_item.annotation.id)
-        
-        if changed_ids:
-            self.selection_changed.emit(changed_ids)
-        
     def mouseReleaseEvent(self, event):
-        """Handle mouse release to complete rubber band selection."""
-        # Check if a rubber band drag was in progress
-        if self.rubber_band_origin is not None and event.button() == Qt.LeftButton:
-            if self.rubber_band and self.rubber_band.isVisible():
-                self.rubber_band.hide()
-                self.rubber_band.deleteLater()
-                self.rubber_band = None
-
-            # **NEEDED CHANGE**: Clean up the stored selection state.
-            self.selection_at_press = set()
+        # This event is now called from the event filter
+        if self.rubber_band and self.rubber_band.isVisible():
+            self.rubber_band.hide()
+            self.rubber_band.deleteLater()
+            self.rubber_band = None
             self.rubber_band_origin = None
-            self.mouse_pressed_on_widget = False
-            event.accept()
             return
-            
         super().mouseReleaseEvent(event)
-        
+
     def keyPressEvent(self, event):
-        """Handle keyboard press events including deletion of selected annotations."""
         if event.key() == Qt.Key_Delete and event.modifiers() == Qt.ControlModifier:
-            # Ctrl+Delete pressed - mark selected annotations for deletion
-            if self.selected_widgets:
-                self.mark_selected_for_deletion()
+            if self.current_selection_ids:
+                self.deletion_requested.emit(list(self.current_selection_ids))
                 event.accept()
                 return
-        
         super().keyPressEvent(event)
 
-    def mark_selected_for_deletion(self):
-            """Mark currently selected annotations for deletion and hide them from view."""
-            if not self.selected_widgets:
-                return
-
-            changed_ids = []
-            widgets_to_hide = []
-            
-            # If in isolation mode, remove the deleted widgets from the isolated set.
-            # This makes the deletion from the isolation view explicit and persistent.
-            if self.isolated_mode:
-                for widget in self.selected_widgets:
-                    if widget in self.isolated_widgets:
-                        self.isolated_widgets.remove(widget)
-
-            for widget in self.selected_widgets:
-                if not widget.data_item.is_marked_for_deletion():
-                    widget.data_item.mark_for_deletion()
-                    widgets_to_hide.append(widget)
-                    changed_ids.append(widget.data_item.annotation.id)
-
-            # Hide the marked widgets from view
-            self.content_widget.setUpdatesEnabled(False)
-            try:
-                for widget in widgets_to_hide:
-                    widget.hide()
-                
-                # Recalculate layout after hiding widgets
-                self.recalculate_widget_positions()
-            finally:
-                self.content_widget.setUpdatesEnabled(True)
-
-            # Clear selection after marking for deletion
-            self.clear_selection()
-
-            if changed_ids:
-                self.preview_changed.emit(changed_ids)
-
-            print(f"Marked {len(changed_ids)} annotation(s) for deletion and hid them from view")
-            
-    def get_visible_widgets(self):
-        """Get widgets that should be visible (not marked for deletion)."""
-        return [w for w in self.annotation_widgets_by_id.values() 
-                if not w.data_item.is_marked_for_deletion()]
-
     def handle_annotation_selection(self, widget, event):
-        """Handle selection of annotation widgets with different modes."""
-        # Get the list of widgets to work with based on isolation mode
-        if self.isolated_mode:
-            # Only work with visible widgets when in isolation mode
-            widget_list = [w for w in self.annotation_widgets_by_id.values() if not w.isHidden()]
-        else:
-            # Use all widgets when not in isolation mode
-            widget_list = list(self.annotation_widgets_by_id.values())
-        
-        try:
-            widget_index = widget_list.index(widget)
-        except ValueError:
-            return
-
+        """INTERNAL: Handles a click on a widget and signals the controller."""
+        target_id = widget.data_item.annotation.id
         modifiers = event.modifiers()
-        changed_ids = []
 
-        # --- The selection logic now identifies which items to change   ---
-        # --- but the core state change happens in select/deselect       ---
-        
-        if modifiers == Qt.ShiftModifier or modifiers == (Qt.ShiftModifier | Qt.ControlModifier):
-            # Range selection
-            if self.last_selected_index != -1:
-                # Find the last selected widget in the current widget list
-                last_selected_widget = None
-                for w in self.selected_widgets:
-                    if w in widget_list:
-                        try:
-                            last_index_in_current_list = widget_list.index(w)
-                            if last_selected_widget is None or \
-                               last_index_in_current_list > widget_list.index(last_selected_widget):
-                                last_selected_widget = w
-                        except ValueError:
-                            continue
-                
-                if last_selected_widget:
-                    last_selected_index_in_current_list = widget_list.index(last_selected_widget)
-                    start = min(last_selected_index_in_current_list, widget_index)
-                    end = max(last_selected_index_in_current_list, widget_index)
-                else:
-                    # Fallback if no previously selected widget is found in current list
-                    start = widget_index
-                    end = widget_index
-                
-                for i in range(start, end + 1):
-                    # select_widget will return True if a change occurred
-                    if self.select_widget(widget_list[i]):
-                        changed_ids.append(widget_list[i].data_item.annotation.id)
-            else:
-                if self.select_widget(widget):
-                    changed_ids.append(widget.data_item.annotation.id)
-                self.last_selected_index = widget_index
-        
+        new_selection = self.current_selection_ids.copy()
+
+        if modifiers == Qt.NoModifier:
+            new_selection = {target_id}
         elif modifiers == Qt.ControlModifier:
-            # Toggle selection
-            if widget.is_selected():
-                if self.deselect_widget(widget):
-                    changed_ids.append(widget.data_item.annotation.id)
-            else:
-                if self.select_widget(widget):
-                    changed_ids.append(widget.data_item.annotation.id)
-            self.last_selected_index = widget_index
-                
-        else:
-            # Normal click: clear all others and select this one
-            newly_selected_id = widget.data_item.annotation.id
-            # Deselect all widgets that are not the clicked one
-            for w in list(self.selected_widgets):
-                if w.data_item.annotation.id != newly_selected_id:
-                    if self.deselect_widget(w):
-                        changed_ids.append(w.data_item.annotation.id)
-            # Select the clicked widget
-            if self.select_widget(widget):
-                changed_ids.append(newly_selected_id)
-            self.last_selected_index = widget_index
-            
-        # Update isolation if in isolated mode
-        if self.isolated_mode:
-            self._update_isolation()
-        
-        # If any selections were changed, emit the signal
-        if changed_ids:
-            self.selection_changed.emit(changed_ids)
-            
-    def _update_isolation(self):
-        """Update the isolated view to show only currently selected widgets."""
-        if not self.isolated_mode:
-            return
-            
-        if self.selected_widgets:
-            self.isolated_widgets.update(self.selected_widgets)
-            self.setUpdatesEnabled(False)
+            new_selection.symmetric_difference_update({target_id})  # Toggle
+        elif modifiers == Qt.ShiftModifier and self.last_selected_id:
+            visible_widgets = self._get_sorted_widgets()
+            ids = [w.data_item.annotation.id for w in visible_widgets]
             try:
-                for widget in self.annotation_widgets_by_id.values():
-                    # A widget must be in the isolated set AND not marked for deletion to be visible.
-                    is_in_isolated_set = widget in self.isolated_widgets
-                    is_marked_for_deletion = widget.data_item.is_marked_for_deletion()
-                    
-                    if is_in_isolated_set and not is_marked_for_deletion:
-                        widget.show()
-                    else:
-                        widget.hide()
-                        
-                self.recalculate_widget_positions()
-            finally:
-                self.setUpdatesEnabled(True)
+                start_idx = ids.index(self.last_selected_id)
+                end_idx = ids.index(target_id)
+                if start_idx > end_idx:
+                    start_idx, end_idx = end_idx, start_idx
+                for i in range(start_idx, end_idx + 1):
+                    new_selection.add(ids[i])
+            except ValueError:
+                new_selection = {target_id}  # Fallback
         else:
-            # If no widgets are selected, keep the current isolation (don't exit)
-            pass
+            new_selection = {target_id}
 
-    def select_widget(self, widget):
-        """Select a widget, update the data_item, and return True if state changed."""
-        if not widget.is_selected():
-            widget.set_selected(True)
-            widget.data_item.set_selected(True)
-            self.selected_widgets.append(widget)
-            self.update_label_window_selection()
-            self._update_toolbar_state()  # Update button states
-            return True
-        return False
+        self.last_selected_id = target_id
+        self.selection_changed.emit(list(new_selection))
 
-    def deselect_widget(self, widget):
-        """Deselect a widget, update the data_item, and return True if state changed."""
-        if widget.is_selected():
-            widget.set_selected(False)
-            widget.data_item.set_selected(False)
-            if widget in self.selected_widgets:
-                self.selected_widgets.remove(widget)
-            self.update_label_window_selection()
-            self._update_toolbar_state()  # Update button states
-            return True
-        return False
-
-    def clear_selection(self):
-        """Clear all selected widgets and update toolbar state."""
-        for widget in list(self.selected_widgets):
-            widget.set_selected(False)
-        self.selected_widgets.clear()
-        self.update_label_window_selection()
-        self._update_toolbar_state()  # Update button states
-
-    def update_label_window_selection(self):
-        """Update the label window selection based on currently selected annotations."""
-        explorer_window = self.parent()
-        while explorer_window and not hasattr(explorer_window, 'main_window'):
-            explorer_window = explorer_window.parent()
-            
-        if not explorer_window or not hasattr(explorer_window, 'main_window'):
-            return
-            
-        main_window = explorer_window.main_window
-        label_window = main_window.label_window
-        annotation_window = main_window.annotation_window
-        
-        if not self.selected_widgets:
-            label_window.deselect_active_label()
-            label_window.update_annotation_count()
-            return
-            
-        selected_data_items = [widget.data_item for widget in self.selected_widgets]
-        
-        first_effective_label = selected_data_items[0].effective_label
-        all_same_current_label = True
-        for item in selected_data_items:
-            if item.effective_label.id != first_effective_label.id:
-                all_same_current_label = False
-                break
-        
-        if all_same_current_label:
-            label_window.set_active_label(first_effective_label)
-            if not selected_data_items[0].has_preview_changes():
-                annotation_window.labelSelected.emit(first_effective_label.id)
-        else:
-            label_window.deselect_active_label()
-        
-        label_window.update_annotation_count()
-
-    def get_selected_annotations(self):
-        """Get the annotations corresponding to selected widgets."""
-        return [widget.annotation for widget in self.selected_widgets]
-    
     def render_selection_from_ids(self, selected_ids):
-        """Update the visual selection of widgets based on a set of IDs from the controller."""
-        # Block signals temporarily to prevent cascade updates
-        self.setUpdatesEnabled(False)
-        
-        try:
-            for ann_id, widget in self.annotation_widgets_by_id.items():
-                is_selected = ann_id in selected_ids
-                widget.set_selected(is_selected)
-            
-            # Resync internal list of selected widgets
-            self.selected_widgets = [w for w in self.annotation_widgets_by_id.values() if w.is_selected()]
-            
-            # If we're in isolated mode, ADD to the isolation instead of replacing it
-            if self.isolated_mode and self.selected_widgets:
-                self.isolated_widgets.update(self.selected_widgets)  # Add to existing isolation
-                # Hide all widgets except those in the isolated set
-                for widget in self.annotation_widgets_by_id.values():
-                    if widget not in self.isolated_widgets:
-                        widget.hide()
-                    else:
-                        widget.show()
-                self.recalculate_widget_positions()
-            
-        finally:
-            self.setUpdatesEnabled(True)
-        
-        # Update label window once at the end
-        self.update_label_window_selection()
-        # Update toolbar state to enable/disable Isolate button
+        """PUBLIC: Updates the visual selection from a set of IDs."""
+        self.current_selection_ids = set(selected_ids)
+        self.content_widget.setUpdatesEnabled(False)
+        for ann_id, widget in self.annotation_widgets_by_id.items():
+            widget.set_selected(ann_id in self.current_selection_ids)
+        self.content_widget.setUpdatesEnabled(True)
+
+        if self.isolated_mode and not self.current_selection_ids.issubset(self.isolated_ids):
+            self.isolated_ids.update(self.current_selection_ids)
+            self.recalculate_widget_positions()
+
         self._update_toolbar_state()
-    
+
     def apply_preview_label_to_selected(self, preview_label):
-        """Apply a preview label and emit a signal for the embedding view to update."""
-        if not self.selected_widgets or not preview_label:
+        """PUBLIC: Applies a preview label and signals the controller."""
+        if not self.current_selection_ids or not preview_label:
             return
 
-        changed_ids = []
-        for widget in self.selected_widgets:
-            widget.data_item.set_preview_label(preview_label)
-            widget.update() # Force repaint with new color
-            changed_ids.append(widget.data_item.annotation.id)
-            
-        # Recalculate positions to update sorting based on new effective labels
+        for ann_id in self.current_selection_ids:
+            if ann_id in self.annotation_widgets_by_id:
+                widget = self.annotation_widgets_by_id[ann_id]
+                widget.data_item.set_preview_label(preview_label)
+                widget.update()
+
         if self.sort_combo.currentText() == "Label":
             self.recalculate_widget_positions()
 
-        if changed_ids:
-            self.preview_changed.emit(changed_ids)
+        self.preview_label_applied.emit()
 
-    def clear_preview_states(self):
-        """Clear all preview states and revert to original labels."""
-        changed_ids = []
-        widgets_to_show = []
-        
+    def refresh_display(self, data_items, selected_ids):
+        """PUBLIC: Re-renders the entire view based on controller's data."""
+        self.update_annotations(data_items)
+        self.render_selection_from_ids(selected_ids)
+
+    def refresh_widget_colors(self):
+        """PUBLIC: Forces all visible widgets to repaint to show new colors."""
+        self.content_widget.setUpdatesEnabled(False)
         for widget in self.annotation_widgets_by_id.values():
-            # Check if this widget's state will be cleared
-            was_preview = widget.data_item.has_preview_changes()
-            was_deleted = widget.data_item.is_marked_for_deletion()
+            if not widget.isHidden():
+                widget.update()
+        self.content_widget.setUpdatesEnabled(True)
 
-            if was_preview:
-                widget.data_item.clear_preview_label()
-                widget.update() # Repaint to show original color
-            
-            if was_deleted:
-                widget.data_item.unmark_for_deletion()
-                widgets_to_show.append(widget)
-            
-            # If the item's state was reverted, add its ID to the list
-            if was_preview or was_deleted:
-                changed_ids.append(widget.data_item.annotation.id)
-
-        # Show any widgets that were previously hidden due to deletion
-        if widgets_to_show:
-            self.content_widget.setUpdatesEnabled(False)
-            try:
-                for widget in widgets_to_show:
-                    if not self.isolated_mode or widget in self.isolated_widgets:
-                        widget.show()
-                self.recalculate_widget_positions()
-            finally:
-                self.content_widget.setUpdatesEnabled(True)
-        
-        # If any changes were made, update UI and notify listeners
-        if changed_ids:
-            if self.sort_combo.currentText() == "Label":
-                self.recalculate_widget_positions()
-            self.update_label_window_selection()
-            self.preview_changed.emit(changed_ids)
-
-    def has_preview_changes(self):
-        """Check if there are any pending preview changes."""
-        return any(w.data_item.has_preview_changes() or w.data_item.is_marked_for_deletion() 
-                  for w in self.annotation_widgets_by_id.values())
-
-    def get_preview_changes_summary(self):
-        """Get a summary of preview changes for user feedback."""
-        label_changes = sum(1 for w in self.annotation_widgets_by_id.values() if w.data_item.has_preview_changes())
-        deletion_marks = sum(1 for w in self.annotation_widgets_by_id.values() if w.data_item.is_marked_for_deletion())
-        
-        if not label_changes and not deletion_marks:
-            return "No preview changes"
-        
-        parts = []
-        if label_changes:
-            parts.append(f"{label_changes} label change(s)")
-        if deletion_marks:
-            parts.append(f"{deletion_marks} deletion(s)")
-        
-        return " and ".join(parts)
-    
-    def apply_preview_changes_permanently(self):
-        """Apply all preview changes permanently to the annotation data."""
-        applied_annotations = []
-        annotations_to_delete = []
-        
-        for widget in self.annotation_widgets_by_id.values():
-            # Handle label changes
-            if widget.data_item.apply_preview_permanently():
-                applied_annotations.append(widget.annotation)
-            
-            # Handle deletion marks
-            if widget.data_item.is_marked_for_deletion():
-                annotations_to_delete.append(widget.annotation)
-        
-        return applied_annotations, annotations_to_delete
-
-
-# ----------------------------------------------------------------------------------------------------------------------
-# Widgets
-# ----------------------------------------------------------------------------------------------------------------------
-
-
-class AnnotationSettingsWidget(QGroupBox):
-    """Widget for filtering annotations by image, type, and label in a multi-column layout."""
-
-    def __init__(self, main_window, parent=None):
-        super(AnnotationSettingsWidget, self).__init__("Annotation Settings", parent)
-        self.main_window = main_window
-        self.explorer_window = parent  # Store reference to ExplorerWindow
-        self.setup_ui()
-
-    def setup_ui(self):
-        # The main layout is vertical, to hold the top columns, the stretch, and the bottom buttons
-        layout = QVBoxLayout(self)
-
-        # A horizontal layout to contain the filter columns
-        conditions_layout = QHBoxLayout()
-
-        # Images column
-        images_column = QVBoxLayout()
-        images_label = QLabel("Images:")
-        images_label.setStyleSheet("font-weight: bold;")
-        images_column.addWidget(images_label)
-
-        self.images_list = QListWidget()
-        self.images_list.setSelectionMode(QListWidget.MultiSelection)
-        self.images_list.setMaximumHeight(50)
-
-        if hasattr(self.main_window, 'image_window') and hasattr(self.main_window.image_window, 'raster_manager'):
-            for path in self.main_window.image_window.raster_manager.image_paths:
-                self.images_list.addItem(os.path.basename(path))
-
-        images_column.addWidget(self.images_list)
-
-        images_buttons_layout = QHBoxLayout()
-        self.images_select_all_btn = QPushButton("Select All")
-        self.images_select_all_btn.clicked.connect(self.select_all_images)
-        images_buttons_layout.addWidget(self.images_select_all_btn)
-
-        self.images_deselect_all_btn = QPushButton("Deselect All")
-        self.images_deselect_all_btn.clicked.connect(self.deselect_all_images)
-        images_buttons_layout.addWidget(self.images_deselect_all_btn)
-        images_column.addLayout(images_buttons_layout)
-
-        conditions_layout.addLayout(images_column)
-
-        # Annotation Type column
-        type_column = QVBoxLayout()
-        type_label = QLabel("Annotation Type:")
-        type_label.setStyleSheet("font-weight: bold;")
-        type_column.addWidget(type_label)
-
-        self.annotation_type_list = QListWidget()
-        self.annotation_type_list.setSelectionMode(QListWidget.MultiSelection)
-        self.annotation_type_list.setMaximumHeight(50)
-        self.annotation_type_list.addItems(["PatchAnnotation",
-                                            "RectangleAnnotation",
-                                            "PolygonAnnotation",
-                                            "MultiPolygonAnnotation"])
-
-        type_column.addWidget(self.annotation_type_list)
-
-        type_buttons_layout = QHBoxLayout()
-        self.type_select_all_btn = QPushButton("Select All")
-        self.type_select_all_btn.clicked.connect(self.select_all_annotation_types)
-        type_buttons_layout.addWidget(self.type_select_all_btn)
-
-        self.type_deselect_all_btn = QPushButton("Deselect All")
-        self.type_deselect_all_btn.clicked.connect(self.deselect_all_annotation_types)
-        type_buttons_layout.addWidget(self.type_deselect_all_btn)
-        type_column.addLayout(type_buttons_layout)
-
-        conditions_layout.addLayout(type_column)
-
-        # Label column
-        label_column = QVBoxLayout()
-        label_label = QLabel("Label:")
-        label_label.setStyleSheet("font-weight: bold;")
-        label_column.addWidget(label_label)
-
-        self.label_list = QListWidget()
-        self.label_list.setSelectionMode(QListWidget.MultiSelection)
-        self.label_list.setMaximumHeight(50)
-
-        if hasattr(self.main_window, 'label_window') and hasattr(self.main_window.label_window, 'labels'):
-            for label in self.main_window.label_window.labels:
-                self.label_list.addItem(label.short_label_code)
-
-        label_column.addWidget(self.label_list)
-
-        label_buttons_layout = QHBoxLayout()
-        self.label_select_all_btn = QPushButton("Select All")
-        self.label_select_all_btn.clicked.connect(self.select_all_labels)
-        label_buttons_layout.addWidget(self.label_select_all_btn)
-
-        self.label_deselect_all_btn = QPushButton("Deselect All")
-        self.label_deselect_all_btn.clicked.connect(self.deselect_all_labels)
-        label_buttons_layout.addWidget(self.label_deselect_all_btn)
-        label_column.addLayout(label_buttons_layout)
-
-        conditions_layout.addLayout(label_column)
-
-        # Add the horizontal layout of columns to the main vertical layout
-        layout.addLayout(conditions_layout)
-
-        # Add a stretch item to push the columns to the top
-        layout.addStretch(1)
-
-        # Bottom buttons layout with Apply and Clear buttons on the right
-        bottom_layout = QHBoxLayout()
-        bottom_layout.addStretch()  # Push buttons to the right
-
-        self.apply_button = QPushButton("Apply")
-        self.apply_button.clicked.connect(self.apply_conditions)
-        bottom_layout.addWidget(self.apply_button)
-
-        self.clear_button = QPushButton("Clear")
-        self.clear_button.clicked.connect(self.clear_all_conditions)
-        bottom_layout.addWidget(self.clear_button)
-
-        # Add the bottom buttons layout to the main layout, keeping it at the bottom
-        layout.addLayout(bottom_layout)
-
-        # Set defaults
-        self.set_defaults()
-
-    def select_all_images(self):
-        """Select all items in the images list."""
-        self.images_list.selectAll()
-
-    def deselect_all_images(self):
-        """Deselect all items in the images list."""
-        self.images_list.clearSelection()
-
-    def select_all_annotation_types(self):
-        """Select all items in the annotation types list."""
-        self.annotation_type_list.selectAll()
-
-    def deselect_all_annotation_types(self):
-        """Deselect all items in the annotation types list."""
-        self.annotation_type_list.clearSelection()
-
-    def select_all_labels(self):
-        """Select all items in the labels list."""
-        self.label_list.selectAll()
-
-    def deselect_all_labels(self):
-        """Deselect all items in the labels list."""
-        self.label_list.clearSelection()
-
-    def set_defaults(self):
-        """Set default selections."""
-        self.set_default_to_current_image()
-        self.select_all_annotation_types()
-        self.select_all_labels()
-
-    def set_default_to_current_image(self):
-        """Set the current image as the default selection."""
-        if hasattr(self.main_window, 'annotation_window'):
-            current_image_path = self.main_window.annotation_window.current_image_path
-            if current_image_path:
-                current_image_name = os.path.basename(current_image_path)
-                items = self.images_list.findItems(current_image_name, Qt.MatchExactly)
-                if items:
-                    items[0].setSelected(True)
-                    return
-        self.select_all_images()
-
-    def clear_all_conditions(self):
-        """Reset all conditions to their defaults."""
-        self.images_list.clearSelection()
-        self.annotation_type_list.clearSelection()
-        self.label_list.clearSelection()
-        self.set_defaults()
-        if self.explorer_window and hasattr(self.explorer_window, 'refresh_filters'):
-            self.explorer_window.refresh_filters()
-
-    def apply_conditions(self):
-        """Apply the current filter conditions."""
-        if self.explorer_window and hasattr(self.explorer_window, 'refresh_filters'):
-            self.explorer_window.refresh_filters()
-
-    def get_selected_images(self):
-        """Get selected image names."""
-        selected_items = self.images_list.selectedItems()
-        if not selected_items:
-            return []
-        return [item.text() for item in selected_items]
-
-    def get_selected_annotation_types(self):
-        """Get selected annotation types."""
-        selected_items = self.annotation_type_list.selectedItems()
-        if not selected_items:
-            return []
-        return [item.text() for item in selected_items]
-
-    def get_selected_labels(self):
-        """Get selected labels."""
-        selected_items = self.label_list.selectedItems()
-        if not selected_items:
-            return []
-        return [item.text() for item in selected_items]
-    
-    
-class ModelSettingsWidget(QGroupBox):
-    """Widget containing model selection with tabs for different model sources."""
-
-    def __init__(self, main_window, parent=None):
-        super(ModelSettingsWidget, self).__init__("Model Settings", parent)
-        self.main_window = main_window
-        self.explorer_window = parent
-        self.setup_ui()
-
-    def setup_ui(self):
-        """Set up the UI with a tabbed interface for model selection."""
-        main_layout = QVBoxLayout(self)
-
-        # --- Tabbed Interface for Model Selection ---
-        self.tabs = QTabWidget()
-
-        # Tab 1: Select Model
-        model_select_tab = QWidget()
-        model_select_layout = QFormLayout(model_select_tab)
-        model_select_layout.setContentsMargins(5, 10, 5, 5) # Add some top margin
-
-        self.model_combo = QComboBox()
-        self.model_combo.addItems(["Color Features"])
-        self.model_combo.insertSeparator(1) # Add a separator
-        
-        standard_models = ['yolov8n-cls.pt',
-                           'yolov8s-cls.pt',
-                           'yolov8m-cls.pt',
-                           'yolov8l-cls.pt',
-                           'yolov8x-cls.pt',
-                           'yolo11n-cls.pt',
-                           'yolo11s-cls.pt',
-                           'yolo11m-cls.pt',
-                           'yolo11l-cls.pt',
-                           'yolo11x-cls.pt',
-                           'yolo12n-cls.pt',
-                           'yolo12s-cls.pt',
-                           'yolo12m-cls.pt',
-                           'yolo12l-cls.pt',
-                           'yolo12x-cls.pt']
-        
-        self.model_combo.addItems(standard_models)
-        
-        community_configs = get_available_configs(task='classify')
-        if community_configs:
-            self.model_combo.insertSeparator(len(standard_models) + 2)
-            self.model_combo.addItems(list(community_configs.keys()))
-            
-        self.model_combo.setCurrentText('Color Features')
-        
-        model_select_layout.addRow("Model:", self.model_combo)
-        self.tabs.addTab(model_select_tab, "Select Model")
-
-        # Tab 2: Existing Model from File
-        model_existing_tab = QWidget()
-        model_existing_layout = QFormLayout(model_existing_tab)
-        model_existing_layout.setContentsMargins(5, 10, 5, 5)
-
-        self.model_path_edit = QLineEdit()
-        self.model_path_edit.setPlaceholderText("Path to a existing .pt model file...")
-        browse_button = QPushButton("Browse...")
-        browse_button.clicked.connect(self.browse_for_model)
-        
-        path_layout = QHBoxLayout()
-        path_layout.addWidget(self.model_path_edit)
-        path_layout.addWidget(browse_button)
-        model_existing_layout.addRow("Model Path:", path_layout)
-        self.tabs.addTab(model_existing_tab, "Use Existing Model")
-        
-        main_layout.addWidget(self.tabs)
-
-    def browse_for_model(self):
-        """Open a file dialog to browse for model files."""
-        options = QFileDialog.Options()
-        file_path, _ = QFileDialog.getOpenFileName(
-            self,
-            "Select Model File",
-            "",
-            "PyTorch Models (*.pt);;All Files (*)",
-            options=options
-        )
-        if file_path:
-            self.model_path_edit.setText(file_path)
-
-    def get_selected_model(self):
-        """Get the currently selected model name or path."""
-        current_tab_index = self.tabs.currentIndex()
-        if current_tab_index == 0:
-            return self.model_combo.currentText()
-        elif current_tab_index == 1:
-            return self.model_path_edit.text()
-        return ""
-    
-    
-class EmbeddingSettingsWidget(QGroupBox):
-    """Widget containing settings with tabs for models and embedding."""
-
-    def __init__(self, main_window, parent=None):
-        super(EmbeddingSettingsWidget, self).__init__("Embedding Settings", parent)
-        self.main_window = main_window
-        self.explorer_window = parent
-        
-        self.setup_ui()
-        
-        # Initial call to set the sliders correctly for the default technique
-        self._update_parameter_sliders()
-
-    def setup_ui(self):
-        """Set up the UI with embedding technique parameters."""
-        main_layout = QVBoxLayout(self)
-        main_layout.setContentsMargins(5, 10, 5, 5)
-
-        # Form layout for embedding settings
-        settings_layout = QFormLayout()
-
-        self.embedding_technique_combo = QComboBox()
-        self.embedding_technique_combo.addItems(["PCA", "TSNE", "UMAP"])
-        self.embedding_technique_combo.currentTextChanged.connect(self._update_parameter_sliders)
-        settings_layout.addRow("Technique:", self.embedding_technique_combo)
-
-        # Slider 1
-        self.param1_label = QLabel("Parameter 1:")
-        param1_layout = QHBoxLayout()
-        self.param1_slider = QSlider(Qt.Horizontal)
-        self.param1_value_label = QLabel("0")
-        self.param1_value_label.setMinimumWidth(25)
-        param1_layout.addWidget(self.param1_slider)
-        param1_layout.addWidget(self.param1_value_label)
-        settings_layout.addRow(self.param1_label, param1_layout)
-        self.param1_slider.valueChanged.connect(
-            lambda v: self.param1_value_label.setText(str(v))
-        )
-
-        # Slider 2
-        self.param2_label = QLabel("Parameter 2:")
-        param2_layout = QHBoxLayout()
-        self.param2_slider = QSlider(Qt.Horizontal)
-        self.param2_value_label = QLabel("0.0")
-        self.param2_value_label.setMinimumWidth(35) # Increased width for larger numbers
-        param2_layout.addWidget(self.param2_slider)
-        param2_layout.addWidget(self.param2_value_label)
-        settings_layout.addRow(self.param2_label, param2_layout)
-
-        self.apply_embedding_button = QPushButton("Apply Embedding")
-        self.apply_embedding_button.clicked.connect(self.apply_embedding)
-        settings_layout.addRow("", self.apply_embedding_button)
-
-        main_layout.addLayout(settings_layout)
-        
-    def _update_parameter_sliders(self):
-        """Enable, disable, and configure sliders based on the selected technique."""
-        technique = self.embedding_technique_combo.currentText()
-
-        # Disconnect any existing connections to prevent conflicts
-        try:
-            self.param2_slider.valueChanged.disconnect()
-        except TypeError:
-            pass # No connection existed
-
-        if technique == "UMAP":
-            # Enable Row 1 for n_neighbors
-            self.param1_label.setEnabled(True)
-            self.param1_slider.setEnabled(True)
-            self.param1_value_label.setEnabled(True)
-            self.param1_label.setText("n_neighbors:")
-            self.param1_slider.setRange(2, 150)
-            self.param1_slider.setValue(15)
-
-            # Enable Row 2 for min_dist
-            self.param2_label.setEnabled(True)
-            self.param2_slider.setEnabled(True)
-            self.param2_value_label.setEnabled(True)
-            self.param2_label.setText("min_dist:")
-            self.param2_slider.setRange(0, 99)
-            self.param2_slider.setValue(10)
-            self.param2_slider.valueChanged.connect(
-                lambda v: self.param2_value_label.setText(f"{v/100.0:.2f}")
-            )
-
-        elif technique == "TSNE":
-            # Enable Row 1 for Perplexity
-            self.param1_label.setEnabled(True)
-            self.param1_slider.setEnabled(True)
-            self.param1_value_label.setEnabled(True)
-            self.param1_label.setText("Perplexity:")
-            self.param1_slider.setRange(5, 50)
-            self.param1_slider.setValue(30)
-            
-            # --- MODIFIED: Enable Row 2 for Early Exaggeration ---
-            self.param2_label.setEnabled(True)
-            self.param2_slider.setEnabled(True)
-            self.param2_value_label.setEnabled(True)
-            self.param2_label.setText("Exaggeration:")
-            self.param2_slider.setRange(50, 600)  # Represents 5.0 to 60.0
-            self.param2_slider.setValue(120)      # Represents 12.0
-            self.param2_slider.valueChanged.connect(
-                lambda v: self.param2_value_label.setText(f"{v/10.0:.1f}")
-            )
-
-        elif technique == "PCA":
-            # Disable both rows for PCA
-            self.param1_label.setEnabled(False)
-            self.param1_slider.setEnabled(False)
-            self.param1_value_label.setEnabled(False)
-            self.param1_label.setText(" ")
-
-            self.param2_label.setEnabled(False)
-            self.param2_slider.setEnabled(False)
-            self.param2_value_label.setEnabled(False)
-            self.param2_label.setText(" ")
-
-    def get_embedding_parameters(self):
-        """Returns a dictionary of the current embedding parameters."""
-        params = {
-            'technique': self.embedding_technique_combo.currentText(),
-        }
-        if params['technique'] == 'UMAP':
-            params['n_neighbors'] = self.param1_slider.value()
-            params['min_dist'] = self.param2_slider.value() / 100.0
-        elif params['technique'] == 'TSNE':
-            params['perplexity'] = self.param1_slider.value()
-            params['early_exaggeration'] = self.param2_slider.value() / 10.0
-        return params
-
-    def apply_embedding(self):
-        if self.explorer_window and hasattr(self.explorer_window, 'run_embedding_pipeline'):
-            # Clear all selections before running embedding pipeline
-            if hasattr(self.explorer_window, 'annotation_viewer'):
-                self.explorer_window.annotation_viewer.clear_selection()
-            if hasattr(self.explorer_window, 'embedding_viewer'):
-                self.explorer_window.embedding_viewer.render_selection_from_ids(set())
-            
-            self.explorer_window.run_embedding_pipeline()
-    
 
 # ----------------------------------------------------------------------------------------------------------------------
 # ExplorerWindow
@@ -2068,821 +789,584 @@ class EmbeddingSettingsWidget(QGroupBox):
 
 
 class ExplorerWindow(QMainWindow):
-    def __init__(self, main_window, parent=None):
-        super(ExplorerWindow, self).__init__(parent)
+    def __init__(self, main_window, label_window=None, parent=None):
+        super().__init__(parent)
         self.main_window = main_window
+        # Use passed widget or fallback
+        self.label_window = label_window or main_window.label_window
+
         self.image_window = main_window.image_window
-        self.label_window = main_window.label_window
         self.annotation_window = main_window.annotation_window
 
-        self.device = main_window.device  # Use the same device as the main window
-        self.model_path = ""
+        # --- Initialize Data Members and State ---
+        self.device = main_window.device
         self.loaded_model = None
-
-        # Store current filtered data items for embedding
-        self.current_data_items = []
-        
-        # Cache for extracted features and the model that generated them ---
+        self.model_path = ""
+        self.data_items_by_id = {}
+        self.current_selection_ids = set()
         self.current_features = None
         self.current_feature_generating_model = ""
 
         self.setWindowTitle("Explorer")
-        # Set the window icon
-        explorer_icon_path = get_icon("magic.png")
-        self.setWindowIcon(QIcon(explorer_icon_path))
+        self.setWindowIcon(get_icon("magic.png"))
 
-        # Create a central widget and main layout
+        # --- Set up the base UI structure ---
         self.central_widget = QWidget()
         self.setCentralWidget(self.central_widget)
         self.main_layout = QVBoxLayout(self.central_widget)
-        # Create a left panel widget and layout for the re-parented LabelWindow
-        self.left_panel = QWidget()
-        self.left_layout = QVBoxLayout(self.left_panel)
 
-        # Create widgets in __init__ so they're always available
+        # --- BUILD THE UI AND CONNECTIONS ONCE ---
+        # These methods will create all child widgets and set up signals.
+        self.setup_ui_widgets()
+        self.setup_connections()
+
+    def setup_ui_widgets(self):
+        """
+        Creates all child widgets and adds them to the layout.
+        This method should only ever be called ONCE from the constructor.
+        """
+        # --- Create all child widgets here ---
         self.annotation_settings_widget = AnnotationSettingsWidget(self.main_window, self)
         self.model_settings_widget = ModelSettingsWidget(self.main_window, self)
         self.embedding_settings_widget = EmbeddingSettingsWidget(self.main_window, self)
-        self.annotation_viewer = AnnotationViewer(self) 
+        self.annotation_viewer = AnnotationViewer(self)
         self.embedding_viewer = EmbeddingViewer(self)
-
-        # Create buttons
         self.clear_preview_button = QPushButton('Clear Preview', self)
-        self.clear_preview_button.clicked.connect(self.clear_preview_changes)
-        self.clear_preview_button.setToolTip("Clear all preview changes and revert to original labels")
-        self.clear_preview_button.setEnabled(False)  # Initially disabled
-
         self.exit_button = QPushButton('Exit', self)
-        self.exit_button.clicked.connect(self.close)
-        self.exit_button.setToolTip("Close the window")
-
         self.apply_button = QPushButton('Apply', self)
-        self.apply_button.clicked.connect(self.apply)
-        self.apply_button.setToolTip("Apply changes")
-        self.apply_button.setEnabled(False)  # Initially disabled
 
-    def showEvent(self, event):
-        self.setup_ui()
-        super(ExplorerWindow, self).showEvent(event)
-
-    def closeEvent(self, event):
-        """
-        Handles the window close event.
-        This now calls the resource cleanup method.
-        """
-        # Stop any running timers to prevent errors
-        if hasattr(self, 'embedding_viewer') and self.embedding_viewer:
-            if hasattr(self.embedding_viewer, 'animation_timer') and self.embedding_viewer.animation_timer:
-                self.embedding_viewer.animation_timer.stop()
-
-        # Clear any unsaved preview states
-        if hasattr(self, 'annotation_viewer'):
-            self.annotation_viewer.clear_preview_states()
-
-        # --- NEW: Call the dedicated cleanup method ---
-        self._cleanup_resources()
-
-        # Re-enable the main window before closing
-        if self.main_window:
-            self.main_window.setEnabled(True)
-
-        # Move the label_window back to the main_window
-        if hasattr(self.main_window, 'explorer_closed'):
-            self.main_window.explorer_closed()
-
-        # Clear the reference in the main_window to allow garbage collection
-        self.main_window.explorer_window = None
-        
-        event.accept()
-
-    def setup_ui(self):
-        # Clear the main layout to remove any existing widgets
-        while self.main_layout.count():
-            child = self.main_layout.takeAt(0)
-            if child.widget():
-                child.widget().setParent(None)  # Remove from layout but don't delete
-
-        # Top section: Conditions and Settings side by side
+        # --- Assemble the layout ---
+        # Top section: Settings
         top_layout = QHBoxLayout()
+        top_layout.addWidget(self.annotation_settings_widget, 2)
+        top_layout.addWidget(self.model_settings_widget, 1)
+        top_layout.addWidget(self.embedding_settings_widget, 1)
+        self.main_layout.addLayout(top_layout)
 
-        # Add existing widgets to layout
-        top_layout.addWidget(self.annotation_settings_widget, 2)  # Give annotation settings more space
-        top_layout.addWidget(self.model_settings_widget, 1)      # Model settings in the middle
-        top_layout.addWidget(self.embedding_settings_widget, 1)  # Embedding settings on the right
-
-        # Create container widget for top layout
-        top_container = QWidget()
-        top_container.setLayout(top_layout)
-        self.main_layout.addWidget(top_container)
-
-        # Middle section: Annotation Viewer (left) and Embedding Viewer (right)
+        # Middle section: Viewers
         middle_splitter = QSplitter(Qt.Horizontal)
-
-        # Wrap annotation viewer in a group box
         annotation_group = QGroupBox("Annotation Viewer")
         annotation_layout = QVBoxLayout(annotation_group)
         annotation_layout.addWidget(self.annotation_viewer)
         middle_splitter.addWidget(annotation_group)
 
-        # Wrap embedding viewer in a group box
         embedding_group = QGroupBox("Embedding Viewer")
         embedding_layout = QVBoxLayout(embedding_group)
         embedding_layout.addWidget(self.embedding_viewer)
         middle_splitter.addWidget(embedding_group)
-
-        # Set splitter proportions (annotation viewer wider)
         middle_splitter.setSizes([700, 300])
-
-        # Add middle section to main layout with stretch factor
         self.main_layout.addWidget(middle_splitter, 1)
 
-        # Note: LabelWindow will be re-parented here by MainWindow.open_explorer_window()
-        # The LabelWindow will be added to self.left_layout at index 1 by the MainWindow
+        # The re-parented Label Window from MainWindow
         self.main_layout.addWidget(self.label_window)
 
-        # Bottom control buttons
-        self.buttons_layout = QHBoxLayout()
-        # Add stretch to push buttons to the right
-        self.buttons_layout.addStretch(1)
+        # Bottom section: Control Buttons
+        buttons_layout = QHBoxLayout()
+        buttons_layout.addStretch(1)
+        buttons_layout.addWidget(self.clear_preview_button)
+        buttons_layout.addWidget(self.exit_button)
+        buttons_layout.addWidget(self.apply_button)
+        self.main_layout.addLayout(buttons_layout)
 
-        # Add existing buttons to layout
-        self.buttons_layout.addWidget(self.clear_preview_button)
-        self.buttons_layout.addWidget(self.exit_button)
-        self.buttons_layout.addWidget(self.apply_button)
-
-        self.main_layout.addLayout(self.buttons_layout)
-
-        # Set default condition to current image and refresh filters
+        # --- Set initial UI state ---
         self.annotation_settings_widget.set_default_to_current_image()
+
+    def showEvent(self, event):
+        """
+        This method now only handles logic that must run *every time* the
+        window becomes visible. It no longer builds or destroys the UI.
+        """
+        super(ExplorerWindow, self).showEvent(event)
+        # Refreshing filters ensures the view is up-to-date with the project state.
         self.refresh_filters()
 
-        # Connect label selection to preview updates (only connect once)
-        try:
-            self.label_window.labelSelected.disconnect(self.on_label_selected_for_preview)
-        except TypeError:
-            pass  # Signal wasn't connected yet
+    def closeEvent(self, event):
+        self._cleanup_resources()
+        if self.main_window:
+            self.main_window.setEnabled(True)
+            if hasattr(self.main_window, 'explorer_closed'):
+                self.main_window.explorer_closed()
+        # This check is important for a clean shutdown
+        if self.main_window:
+            self.main_window.explorer_window = None
+        event.accept()
 
-        self.label_window.labelSelected.connect(self.on_label_selected_for_preview)
-        self.annotation_viewer.selection_changed.connect(self.on_annotation_view_selection_changed)
-        self.annotation_viewer.preview_changed.connect(self.on_preview_changed)
-        self.annotation_viewer.reset_view_requested.connect(self.on_reset_view_requested)
-        self.embedding_viewer.selection_changed.connect(self.on_embedding_view_selection_changed)
-        self.embedding_viewer.reset_view_requested.connect(self.on_reset_view_requested)
+    def setup_connections(self):
+        """Central place to connect all signals and slots for the Controller pattern."""
+        # --- Controller actions from buttons ---
+        self.clear_preview_button.clicked.connect(self.clear_all_previews)
+        self.exit_button.clicked.connect(self.close)
+        self.apply_button.clicked.connect(self.apply_all_changes)
 
-    @pyqtSlot(list)
-    def on_annotation_view_selection_changed(self, changed_ann_ids):
-        """A selection was made in the AnnotationViewer, so update the EmbeddingViewer."""
-        all_selected_ids = {w.data_item.annotation.id for w in self.annotation_viewer.selected_widgets}
-        
-        # Only try to sync the selection with the EmbeddingViewer if it has points.
-        # This prevents the feedback loop that was clearing the selection.
-        if self.embedding_viewer.points_by_id:
-            self.embedding_viewer.render_selection_from_ids(all_selected_ids)
-            
-        self.update_label_window_selection() # Keep label window in sync
+        # --- Viewers -> Controller signals ---
+        self.annotation_viewer.selection_changed.connect(self.handle_selection_change)
+        self.annotation_viewer.deletion_requested.connect(self.handle_deletion_request)
+        self.annotation_viewer.preview_label_applied.connect(self.handle_preview_update)
+        self.annotation_viewer.reset_view_requested.connect(self.handle_view_reset)
 
-    @pyqtSlot(list)
-    def on_embedding_view_selection_changed(self, all_selected_ann_ids):
-        """A selection was made in the EmbeddingViewer, so update the AnnotationViewer."""
-        # Check if this is a new selection being made when nothing was previously selected
-        was_empty_selection = len(self.annotation_viewer.selected_widgets) == 0
-        is_new_selection = len(all_selected_ann_ids) > 0
-        
-        # Update the annotation viewer with the new selection
-        self.annotation_viewer.render_selection_from_ids(set(all_selected_ann_ids))
-        
-        # Auto-switch to isolation mode if conditions are met
-        if (was_empty_selection and 
-            is_new_selection and 
-            not self.annotation_viewer.isolated_mode):
-            print("Auto-switching to isolation mode due to new selection in embedding viewer")
-            self.annotation_viewer.isolate_selection()
-        
-        self.update_label_window_selection() # Keep label window in sync
+        self.embedding_viewer.selection_changed.connect(self.handle_selection_change)
+        self.embedding_viewer.reset_view_requested.connect(self.handle_view_reset)
+
+        # --- Other UI elements -> Controller signals ---
+        self.label_window.labelSelected.connect(self.handle_label_selection_for_preview)
+
+    # --- SLOTS: Handling signals from viewers and UI ---
 
     @pyqtSlot(list)
-    def on_preview_changed(self, changed_ann_ids):
-        """A preview color OR deletion state was changed, so update the EmbeddingViewer points."""
-        for ann_id in changed_ann_ids:
-            point = self.embedding_viewer.points_by_id.get(ann_id)
-            if point:
-                data_item = point.data(0)
-                if data_item:
-                    # Set point visibility based on the data item's deletion status
-                    point.setVisible(not data_item.is_marked_for_deletion())
-                    # Force a repaint to apply visibility and any color changes
-                    point.update()
-                
-    @pyqtSlot()
-    def on_reset_view_requested(self):
-        """Handle reset view requests from double-click in either viewer."""
-        # Clear all selections in both viewers
-        self.annotation_viewer.clear_selection()
-        self.embedding_viewer.render_selection_from_ids(set())
-        
-        # Exit isolation mode if currently active
-        if self.annotation_viewer.isolated_mode:
-            self.annotation_viewer.show_all_annotations()
-        
-        # Update button states
+    def handle_selection_change(self, selected_ids):
+        """Handles selection changes from ANY viewer, acting as the single point of truth."""
+        new_selection = set(selected_ids)
+        if new_selection == self.current_selection_ids:
+            return  # No change
+
+        self.current_selection_ids = new_selection
+
+        # Update the central data model's selection state
+        for item in self.data_items_by_id.values():
+            item.set_selected(item.annotation.id in self.current_selection_ids)
+
+        # Command both viewers to update their visual selection
+        self.annotation_viewer.render_selection_from_ids(self.current_selection_ids)
+        self.embedding_viewer.render_selection_from_ids(self.current_selection_ids)
+
+        # Update other UI elements that depend on selection
+        self._update_label_window_for_selection()
+
+    @pyqtSlot(list)
+    def handle_deletion_request(self, ids_to_delete):
+        """Marks items for deletion and refreshes all views."""
+        for ann_id in ids_to_delete:
+            if ann_id in self.data_items_by_id:
+                self.data_items_by_id[ann_id].mark_for_deletion()
+
+        # Since items were deleted, the selection is now invalid. Clear it.
+        self.handle_selection_change([])
+
+        # Refresh displays to hide the marked items
+        self._refresh_all_views()
         self.update_button_states()
-        
-        print("Reset view: cleared selections and exited isolation mode")
 
-    def update_label_window_selection(self):
-        """Update the label window based on the selection in the annotation viewer."""
-        self.annotation_viewer.update_label_window_selection()
+    @pyqtSlot()
+    def handle_preview_update(self):
+        """Handles updates after a preview label is applied to the selection."""
+        # Command viewers to update visuals that depend on data item state
+        self.embedding_viewer.refresh_point_colors()
+        self.annotation_viewer.refresh_widget_colors()  # Use this instead of full repaint
 
-    def get_filtered_data_items(self):
-        """Get annotations that match all conditions, returned as AnnotationDataItem objects."""
-        data_items = []
+        # Update other UI
+        self._update_label_window_for_selection()
+        self.update_button_states()
+
+    @pyqtSlot(object)  # Can be Label or LabelScheme
+    def handle_label_selection_for_preview(self, label):
+        """Applies a selected label from the LabelWindow as a preview to the current selection."""
+        if not self.current_selection_ids:
+            return
+
+        for ann_id in self.current_selection_ids:
+            if ann_id in self.data_items_by_id:
+                self.data_items_by_id[ann_id].set_preview_label(label)
+
+        # After changing the data, signal that a preview has updated
+        self.handle_preview_update()
+
+    @pyqtSlot()
+    def handle_view_reset(self):
+        """Resets selection and isolated view, requested from any viewer's double-click."""
+        self.annotation_viewer.show_all_annotations()
+        self.handle_selection_change([])
+
+    # --- PUBLIC METHODS: High-level user actions ---
+
+    def refresh_filters(self):
+        """Filters annotations and completely refreshes all views and data."""
+        QApplication.setOverrideCursor(Qt.WaitCursor)
+        try:
+            # 1. Get filtered annotations and create the central data items
+            filtered_annotations = self._get_filtered_annotations()
+            self._ensure_cropped_images(filtered_annotations)
+
+            self.data_items_by_id = {ann.id: AnnotationDataItem(ann) for ann in filtered_annotations}
+
+            # 2. Reset runtime state
+            self.current_selection_ids.clear()
+            self.current_features = None  # Invalidate feature cache
+
+            # 3. Command all visual components to refresh
+            self._refresh_all_views()
+            self.embedding_viewer.show_placeholder()  # Embedding is now out of date
+            self.update_button_states()
+
+        finally:
+            QApplication.restoreOverrideCursor()
+
+    def run_embedding_pipeline(self):
+        """Orchestrates the feature extraction and dimensionality reduction pipeline."""
+        if not self.data_items_by_id:
+            return
+
+        embedding_params = self.embedding_settings_widget.get_embedding_parameters()
+        selected_model = self.model_settings_widget.get_selected_model()
+
+        progress_bar = ProgressBar(self, "Generating Embedding")
+        progress_bar.show()
+        QApplication.setOverrideCursor(Qt.WaitCursor)
+
+        try:
+            # Step 1: Feature Extraction (if needed)
+            if self.current_features is None or selected_model != self.current_feature_generating_model:
+                visible_data_items = []
+                for item in self.data_items_by_id.values():
+                    if not item.is_marked_for_deletion():
+                        visible_data_items.append(item)
+
+                features, valid_items = self._extract_features(visible_data_items, progress_bar)
+
+                if len(valid_items) != len(visible_data_items):
+                    valid_ids = {item.annotation.id for item in valid_items}
+                    # Remove items not in valid_ids unless marked for deletion
+                    new_data_items_by_id = {}
+                    for id, item in self.data_items_by_id.items():
+                        if id in valid_ids or item.is_marked_for_deletion():
+                            new_data_items_by_id[id] = item
+
+                    self.data_items_by_id = new_data_items_by_id
+                    self._refresh_all_views()
+
+                self.current_features = features
+                self.current_feature_generating_model = selected_model
+            else:
+                features = self.current_features
+
+            if features is None or len(features) == 0:
+                return
+
+            # Step 2: Dimensionality Reduction
+            progress_bar.set_busy_mode(f"Running {embedding_params['technique']}...")
+            embedded_features = self._run_dimensionality_reduction(features, embedding_params)
+            if embedded_features is None:
+                return
+
+            # Step 3: Update and Visualize
+            progress_bar.set_busy_mode("Updating visualization...")
+            visible_data_items = []
+            for item in self.data_items_by_id.values():
+                if not item.is_marked_for_deletion():
+                    visible_data_items.append(item)
+
+            self._update_data_items_with_embedding(visible_data_items, embedded_features)
+
+            self.embedding_viewer.refresh_display(visible_data_items, self.current_selection_ids)
+            self.embedding_viewer.fit_view_to_points()
+
+        finally:
+            QApplication.restoreOverrideCursor()
+            progress_bar.finish_progress()
+            progress_bar.close()
+
+    def clear_all_previews(self):
+        """Clears all preview labels and un-marks all deletions."""
+        for item in self.data_items_by_id.values():
+            item.clear_preview_label()
+            item.unmark_for_deletion()
+
+        self._refresh_all_views()
+        self.update_button_states()
+        self._update_label_window_for_selection()
+
+    def apply_all_changes(self):
+        """Applies all preview changes and deletions permanently."""
+        QApplication.setOverrideCursor(Qt.WaitCursor)
+        try:
+            applied_annotations = []
+            annotations_to_delete = []
+
+            for item in self.data_items_by_id.values():
+                if item.apply_preview_permanently():
+                    applied_annotations.append(item.annotation)
+                if item.is_marked_for_deletion():
+                    annotations_to_delete.append(item.annotation)
+
+            if applied_annotations:
+                affected_images = {ann.image_path for ann in applied_annotations}
+                for image_path in affected_images:
+                    self.image_window.update_image_annotations(image_path)
+                print(f"Applied label changes to {len(applied_annotations)} annotation(s)")
+
+            if annotations_to_delete:
+                self.annotation_window.delete_annotations(annotations_to_delete)
+                print(f"Deleted {len(annotations_to_delete)} annotation(s)")
+
+            if applied_annotations or annotations_to_delete:
+                self.annotation_window.load_annotations()
+                self.refresh_filters()
+
+            self.update_button_states()
+        finally:
+            QApplication.restoreOverrideCursor()
+
+    # --- HELPER / PRIVATE METHODS ---
+
+    def _refresh_all_views(self):
+        """Commands all viewers to refresh based on the current state of the data model."""
+        visible_items = [item for item in self.data_items_by_id.values() if not item.is_marked_for_deletion()]
+
+        self.annotation_viewer.refresh_display(list(self.data_items_by_id.values()), self.current_selection_ids)
+        self.embedding_viewer.refresh_display(visible_items, self.current_selection_ids)
+
+    def _update_label_window_for_selection(self):
+        """Updates the LabelWindow based on the current selection's labels."""
+        if not self.current_selection_ids:
+            self.label_window.deselect_active_label()
+            self.label_window.update_annotation_count()
+            return
+
+        selected_items = []
+        for ann_id in self.current_selection_ids:
+            if ann_id in self.data_items_by_id:
+                selected_items.append(self.data_items_by_id[ann_id])
+
+        if not selected_items:
+            return
+
+        first_label = selected_items[0].effective_label
+        all_same = all(item.effective_label.id == first_label.id for item in selected_items)
+
+        if all_same:
+            self.label_window.set_active_label(first_label)
+        else:
+            self.label_window.deselect_active_label()
+
+        self.label_window.update_annotation_count()
+
+    def update_button_states(self):
+        """Enables/disables the 'Apply' and 'Clear' buttons based on pending changes."""
+        has_preview_changes = any(item.has_preview_changes() for item in self.data_items_by_id.values())
+        has_deletions = any(item.is_marked_for_deletion() for item in self.data_items_by_id.values())
+        has_changes = has_preview_changes or has_deletions
+        self.clear_preview_button.setEnabled(has_changes)
+        self.apply_button.setEnabled(has_changes)
+
+    def _get_filtered_annotations(self):
+        """Gets annotations from AnnotationWindow based on filter widget settings."""
         if not hasattr(self.main_window.annotation_window, 'annotations_dict'):
-            return data_items
+            return []
 
-        # Get current filter conditions
         selected_images = self.annotation_settings_widget.get_selected_images()
         selected_types = self.annotation_settings_widget.get_selected_annotation_types()
         selected_labels = self.annotation_settings_widget.get_selected_labels()
 
+        if not all([selected_images, selected_types, selected_labels]):
+            return []
+
         annotations_to_process = []
-        for annotation in self.main_window.annotation_window.annotations_dict.values():
-            annotation_matches = True
+        for ann in self.main_window.annotation_window.annotations_dict.values():
+            if (os.path.basename(ann.image_path) in selected_images and
+                type(ann).__name__ in selected_types and
+                    ann.label.short_label_code in selected_labels):
+                annotations_to_process.append(ann)
 
-            # Check image condition - if empty list, no annotations match
-            if selected_images:
-                annotation_image = os.path.basename(annotation.image_path)
-                if annotation_image not in selected_images:
-                    annotation_matches = False
-            else:
-                # No images selected means no annotations should match
-                annotation_matches = False
+        return annotations_to_process
 
-            # Check annotation type condition - if empty list, no annotations match
-            if annotation_matches:
-                if selected_types:
-                    annotation_type = type(annotation).__name__
-                    if annotation_type not in selected_types:
-                        annotation_matches = False
-                else:
-                    # No types selected means no annotations should match
-                    annotation_matches = False
-
-            # Check label condition - if empty list, no annotations match
-            if annotation_matches:
-                if selected_labels:
-                    annotation_label = annotation.label.short_label_code
-                    if annotation_label not in selected_labels:
-                        annotation_matches = False
-                else:
-                    # No labels selected means no annotations should match
-                    annotation_matches = False
-
-            if annotation_matches:
-                annotations_to_process.append(annotation)
-
-        # Ensure all filtered annotations have cropped images
-        self._ensure_cropped_images(annotations_to_process)
-        
-        # Wrap in AnnotationDataItem
-        for ann in annotations_to_process:
-            data_items.append(AnnotationDataItem(ann))
-
-        return data_items
-    
     def _ensure_cropped_images(self, annotations):
         """Ensure all provided annotations have a cropped image available."""
         annotations_by_image = {}
         for annotation in annotations:
-            # Only process annotations that don't have a cropped image yet
             if not annotation.cropped_image:
                 image_path = annotation.image_path
                 if image_path not in annotations_by_image:
                     annotations_by_image[image_path] = []
                 annotations_by_image[image_path].append(annotation)
-        
-        # Only proceed if there are annotations that actually need cropping
+
         if annotations_by_image:
             progress_bar = ProgressBar(self, "Cropping Image Annotations")
             progress_bar.show()
             progress_bar.start_progress(len(annotations_by_image))
-            
+
             try:
-                # Crop annotations for each image using the AnnotationWindow method
-                # This ensures consistency with how cropped images are generated elsewhere
                 for image_path, image_annotations in annotations_by_image.items():
                     self.annotation_window.crop_annotations(
-                        image_path=image_path, 
-                        annotations=image_annotations, 
-                        return_annotations=False,  # We don't need the return value
+                        image_path=image_path,
+                        annotations=image_annotations,
+                        return_annotations=False,
                         verbose=False
                     )
-                    # Update progress bar
                     progress_bar.update_progress()
-                    
-            except Exception as e:
-                print(f"Error cropping annotations: {e}")
-                
             finally:
                 progress_bar.finish_progress()
-                progress_bar.stop_progress()
                 progress_bar.close()
 
-    def _extract_color_features(self, data_items, progress_bar=None, bins=32):
-        """
-        Extracts a comprehensive set of color features from annotation crops using only NumPy.
-        
-        For each image, it calculates:
-        1.  Color Moments (per channel):
-            - Mean (1st moment)
-            - Standard Deviation (2nd moment)
-            - Skewness (3rd moment)
-            - Kurtosis (4th moment)
-        2.  Color Histogram (per channel)
-        3.  Grayscale Statistics:
-            - Mean Brightness
-            - Contrast (Std Dev)
-            - Intensity Range
-        4.  Geometric Features:
-            - Area
-            - Perimeter
-        """
-        if progress_bar:
-            progress_bar.set_title("Extracting Color Features...")
-            progress_bar.start_progress(len(data_items))
-
-        features = []
-        valid_data_items = []
-        for item in data_items:
-            pixmap = item.annotation.get_cropped_image()
-            if pixmap and not pixmap.isNull():
-                # arr has shape (height, width, 3)
-                arr = pixmap_to_numpy(pixmap)
-                
-                # Reshape for channel-wise statistics: (num_pixels, 3)
-                pixels = arr.reshape(-1, 3)
-
-                # --- 1. Calculate Color Moments using only NumPy ---
-                mean_color = np.mean(pixels, axis=0)
-                std_color = np.std(pixels, axis=0)
-                
-                # Center the data (subtract the mean) for skew/kurtosis calculation
-                centered_pixels = pixels - mean_color
-                
-                # A small value (epsilon) is added to the denominator to prevent division by zero
-                epsilon = 1e-8
-                skew_color = np.mean(centered_pixels**3, axis=0) / (std_color**3 + epsilon)
-                kurt_color = np.mean(centered_pixels**4, axis=0) / (std_color**4 + epsilon) - 3
-                
-                # --- 2. Calculate Color Histograms ---
-                histograms = []
-                for i in range(3): # For each channel (R, G, B)
-                    hist, _ = np.histogram(pixels[:, i], bins=bins, range=(0, 255))
-                    
-                    # Normalize histogram
-                    hist_sum = np.sum(hist)
-                    if hist_sum > 0:
-                        histograms.append(hist / hist_sum)
-                    else: # Avoid division by zero
-                        histograms.append(np.zeros(bins))
-                
-                # --- 3. Calculate Grayscale Statistics ---
-                gray_arr = np.dot(arr[..., :3], [0.2989, 0.5870, 0.1140])
-                grayscale_stats = np.array([
-                    np.mean(gray_arr),  # Overall brightness
-                    np.std(gray_arr),   # Overall contrast
-                    np.ptp(gray_arr)    # Peak-to-peak intensity range
-                ])
-                
-                # --- 4. Calculate Geometric Features ---
-                area = getattr(item.annotation, 'area', 0.0)
-                perimeter = getattr(item.annotation, 'perimeter', 0.0)
-                geometric_features = np.array([area, perimeter])
-                
-                # --- 5. Concatenate all features into a single vector ---
-                current_features = np.concatenate([
-                    mean_color,
-                    std_color,
-                    skew_color,
-                    kurt_color,
-                    *histograms,
-                    grayscale_stats,
-                    geometric_features
-                ])
-                
-                features.append(current_features)
-                valid_data_items.append(item)
-            else:
-                print(f"Warning: Could not get cropped image for annotation ID {item.annotation.id}. Skipping.")
-            
-            if progress_bar:
-                progress_bar.update_progress()
-
-        return np.array(features), valid_data_items
-
-    def _extract_yolo_features(self, data_items, model_name, progress_bar=None):
-        """
-        Extracts features from annotation crops using a specified YOLO model.
-        This version processes images as a stream for better memory efficiency.
-        """
-        # Load or retrieve the cached model
-        if model_name != self.model_path or self.loaded_model is None:
-            try:
-                print(f"Loading new model: {model_name}")
-                self.loaded_model = YOLO(model_name)
-                self.model_path = model_name
-                
-                # Determine image size from model config if possible
-                try:
-                    self.imgsz = self.loaded_model.model.args['imgsz']
-                    if self.imgsz > 224:
-                        self.imgsz = 128
-                except (AttributeError, KeyError):
-                    self.imgsz = 128
-                
-                # Run a dummy inference to warm up the model
-                print(f"Warming up model on device '{self.device}'...")
-                dummy_image = np.zeros((self.imgsz, self.imgsz, 3), dtype=np.uint8)
-                self.loaded_model.embed(dummy_image, imgsz=self.imgsz, half=True, device=self.device, verbose=False)
-                    
-            except Exception as e:
-                print(f"ERROR: Could not load YOLO model '{model_name}': {e}")
-                return np.array([]), []
-                
-        if progress_bar:
-            progress_bar.set_title(f"Preparing images...")
-            progress_bar.start_progress(len(data_items))
-
-        # --- RE-IMPLEMENTED: Two-stage process for streaming ---
-        
-        # 1. Prepare a list of all valid images and their corresponding data items.
-        image_list = []
-        valid_data_items = []
-        for item in data_items:
-            pixmap = item.annotation.get_cropped_image()
-            if pixmap and not pixmap.isNull():
-                image_np = pixmap_to_numpy(pixmap)
-                image_list.append(image_np)
-                valid_data_items.append(item)
-            else:
-                print(f"Warning: Could not get cropped image for annotation ID {item.annotation.id}. Skipping.")
-
-            if progress_bar:
-                progress_bar.update_progress()
-
-        if not valid_data_items:
-            print("Warning: No valid images found to process.")
-            return np.array([]), []
-
-        embeddings_list = []
-        
-        try:
-            if progress_bar:
-                progress_bar.set_busy_mode(f"Extracting features with {os.path.basename(model_name)}...")
-            
-            # 2. Pass the entire list of valid images to the model with stream=True.
-            # This returns a generator that yields results one by one.
-            results_generator = self.loaded_model.embed(
-                image_list, 
-                stream=True, 
-                imgsz=self.imgsz, 
-                half=True, 
-                device=self.device, 
-                verbose=False
-            )
-            
-            if progress_bar:
-                progress_bar.set_title(f"Extracting features with {os.path.basename(model_name)}...")
-                progress_bar.start_progress(len(data_items))
-            
-            # 3. Process the results from the generator.
-            for embedding_result in results_generator:
-                embeddings_list.append(embedding_result.cpu().numpy())
-                # Update progress for each item as it's processed from the stream.
-                if progress_bar:
-                    progress_bar.update_progress()
-
-            if not embeddings_list:
-                print("Warning: No features were extracted. The model may have failed.")
-                return np.array([]), []
-
-            embeddings = np.array(embeddings_list)
-            
-        except Exception as e:
-            print(f"ERROR: An error occurred during feature extraction: {e}")
-            return np.array([]), []
-            
-        finally:
-            # Clean up CUDA memory after the operation
-            if torch.cuda.is_available():
-                torch.cuda.empty_cache()
-        
-        print(f"Successfully extracted {len(embeddings)} features.")
-        return embeddings, valid_data_items
-
     def _extract_features(self, data_items, progress_bar=None):
-        """
-        Dispatcher method to call the appropriate feature extraction function.
-        It now passes the progress_bar object to the sub-methods.
-        """
+        """Dispatcher method to call the appropriate feature extraction function."""
         model_name = self.model_settings_widget.get_selected_model()
-
         if not model_name:
-            print("No model selected or path provided.")
             return np.array([]), []
 
-        # --- MODIFIED: Pass the progress_bar object ---
         if model_name == "Color Features":
             return self._extract_color_features(data_items, progress_bar=progress_bar)
         elif ".pt" in model_name:
             return self._extract_yolo_features(data_items, model_name, progress_bar=progress_bar)
         else:
-            print(f"Unknown or invalid feature model selected: {model_name}")
             return np.array([]), []
 
+    def _extract_color_features(self, data_items, progress_bar=None, bins=32):
+        if progress_bar:
+            progress_bar.set_title("Extracting Color Features...")
+            progress_bar.start_progress(len(data_items))
+
+        features, valid_data_items = [], []
+        for item in data_items:
+            pixmap = item.annotation.get_cropped_image()
+            if pixmap and not pixmap.isNull():
+                # Convert QPixmap to numpy array (H, W, 3)
+                arr = pixmap_to_numpy(pixmap)
+                # Flatten image to (N, 3) for RGB pixel stats
+                pixels = arr.reshape(-1, 3)
+                # Mean and std for each channel
+                mean_color, std_color = np.mean(pixels, axis=0), np.std(pixels, axis=0)
+                # Small value to avoid division by zero
+                epsilon = 1e-8
+                # Center pixels for skew/kurtosis
+                centered = pixels - mean_color
+                # Skewness for each channel
+                skew = np.mean(centered**3, axis=0) / (std_color**3 + epsilon)
+                # Kurtosis for each channel
+                kurt = np.mean(centered**4, axis=0) / (std_color**4 + epsilon) - 3
+                # Histograms for R,G,B
+                hists = [np.histogram(pixels[:, i], bins=bins, range=(0, 255))[0] for i in range(3)]
+                # Normalize histograms
+                hists_norm = [h / (h.sum() + epsilon) for h in hists]
+                # Convert to grayscale
+                gray_arr = np.dot(arr[..., :3], [0.2989, 0.5870, 0.1140])
+                # Mean, std, range of gray
+                gray_stats = np.array([np.mean(gray_arr), np.std(gray_arr), np.ptp(gray_arr)])
+                # Area feature (if available), Perimeter feature (if available)
+                geom_feats = np.array([getattr(item.annotation, 'area', 0.0),
+                                       getattr(item.annotation, 'perimeter', 0.0)])
+
+                current_features = np.concatenate([mean_color,
+                                                   std_color,
+                                                   skew,
+                                                   kurt,
+                                                   *hists_norm,
+                                                   gray_stats,
+                                                   geom_feats])
+                features.append(current_features)
+                valid_data_items.append(item)
+
+            if progress_bar:
+                progress_bar.update_progress()
+        return np.array(features), valid_data_items
+
+    def _extract_yolo_features(self, data_items, model_name, progress_bar=None):
+        if model_name != self.model_path or self.loaded_model is None:
+            try:
+                self.loaded_model = YOLO(model_name)
+                self.model_path = model_name
+                self.imgsz = self.loaded_model.model.args.get('imgsz', 128)
+                if self.imgsz > 224:
+                    self.imgsz = 128
+                dummy_image = np.zeros((self.imgsz, self.imgsz, 3), dtype=np.uint8)
+                self.loaded_model.embed(dummy_image, imgsz=self.imgsz, half=True, device=self.device, verbose=False)
+            except Exception as e:
+                print(f"ERROR: Could not load YOLO model '{model_name}': {e}")
+                return np.array([]), []
+
+        if progress_bar:
+            progress_bar.set_title("Preparing images...")
+            progress_bar.start_progress(len(data_items))
+
+        image_list, valid_data_items = [], []
+        for item in data_items:
+            pixmap = item.annotation.get_cropped_image()
+            if pixmap and not pixmap.isNull():
+                image_list.append(pixmap_to_numpy(pixmap))
+                valid_data_items.append(item)
+            if progress_bar:
+                progress_bar.update_progress()
+        if not valid_data_items:
+            return np.array([]), []
+
+        embeddings_list = []
+        try:
+            if progress_bar:
+                progress_bar.set_busy_mode(f"Extracting features...")
+            results_gen = self.loaded_model.embed(image_list,
+                                                  stream=True,
+                                                  imgsz=self.imgsz,
+                                                  half=True,
+                                                  device=self.device,
+                                                  verbose=False)
+            if progress_bar:
+                progress_bar.start_progress(len(valid_data_items))
+
+            for result in results_gen:
+                embeddings_list.append(result.cpu().numpy())
+                if progress_bar:
+                    progress_bar.update_progress()
+
+        finally:
+            if torch.cuda.is_available():
+                torch.cuda.empty_cache()
+        return np.array(embeddings_list), valid_data_items
+
     def _run_dimensionality_reduction(self, features, params):
-        """
-        Runs PCA, UMAP or t-SNE on the feature matrix using provided parameters.
-        """
-        technique = params.get('technique', 'UMAP') # Changed default to UMAP
-        random_state = 42
-        
-        print(f"Running {technique} on {len(features)} items with params: {params}")
-        if len(features) <= 2: # UMAP/t-SNE need at least a few points
-            print("Not enough data points for dimensionality reduction.")
+        technique = params.get('technique', 'UMAP')
+        if len(features) <= 2:
             return None
 
         try:
-            # Scaling is crucial, your implementation is already correct
-            scaler = StandardScaler()
-            features_scaled = scaler.fit_transform(features)
-            
+            features_scaled = StandardScaler().fit_transform(features)
+
             if technique == "UMAP":
-                # Get hyperparameters from params, with sensible defaults
-                n_neighbors = params.get('n_neighbors', 15)
-                min_dist = params.get('min_dist', 0.1)
-                metric = params.get('metric', 'cosine')  # Allow metric to be specified
+                reducer = UMAP(n_components=2,
+                               random_state=42,
+                               n_neighbors=min(params.get('n_neighbors', 15), len(features_scaled) - 1),
+                               min_dist=params.get('min_dist', 0.1), metric='cosine')
 
-                reducer = UMAP(
-                    n_components=2, 
-                    random_state=random_state, 
-                    n_neighbors=min(n_neighbors, len(features_scaled) - 1),
-                    min_dist=min_dist,
-                    metric=metric  # Use the specified metric
-                )
-            
             elif technique == "TSNE":
-                perplexity = params.get('perplexity', 30)
-                early_exaggeration = params.get('early_exaggeration', 12.0)
-                learning_rate = params.get('learning_rate', 'auto')
+                reducer = TSNE(n_components=2,
+                               random_state=42,
+                               perplexity=min(params.get('perplexity', 30), len(features_scaled) - 1),
+                               early_exaggeration=params.get('early_exaggeration', 12.0),
+                               learning_rate='auto', init='pca')
 
-                reducer = TSNE(
-                    n_components=2, 
-                    random_state=random_state, 
-                    perplexity=min(perplexity, len(features_scaled) - 1),
-                    early_exaggeration=early_exaggeration,
-                    learning_rate=learning_rate,
-                    init='pca'  # Improves stability and speed
-                )
-                
             elif technique == "PCA":
-                reducer = PCA(n_components=2, random_state=random_state)
+                reducer = PCA(n_components=2, random_state=42)
 
             else:
-                print(f"Unknown dimensionality reduction technique: {technique}")
                 return None
-                
+
             return reducer.fit_transform(features_scaled)
-            
+
         except Exception as e:
             print(f"Error during {technique} dimensionality reduction: {e}")
             return None
 
     def _update_data_items_with_embedding(self, data_items, embedded_features):
-        """Updates AnnotationDataItem objects with embedding results."""
         scale_factor = 4000
         min_vals = np.min(embedded_features, axis=0)
         max_vals = np.max(embedded_features, axis=0)
-        range_vals = max_vals - min_vals
+        # Add epsilon to avoid division by zero
+        range_vals = max_vals - min_vals + 1e-8
 
         for i, item in enumerate(data_items):
-            # Normalize coordinates for consistent display
-            norm_x = (embedded_features[i, 0] - min_vals[0]) / range_vals[0] if range_vals[0] > 0 else 0.5
-            norm_y = (embedded_features[i, 1] - min_vals[1]) / range_vals[1] if range_vals[1] > 0 else 0.5
-            # Scale and center the points in the view
+            norm_x = (embedded_features[i, 0] - min_vals[0]) / range_vals[0]
+            norm_y = (embedded_features[i, 1] - min_vals[1]) / range_vals[1]
             item.embedding_x = (norm_x * scale_factor) - (scale_factor / 2)
             item.embedding_y = (norm_y * scale_factor) - (scale_factor / 2)
 
-    def run_embedding_pipeline(self):
-        """
-        Orchestrates the feature extraction and dimensionality reduction pipeline.
-        This version correctly re-runs reduction on cached features when parameters change.
-        """
-        if not self.current_data_items:
-            print("No items to process for embedding.")
-            return
-
-        # 1. Get current parameters from the UI
-        embedding_params = self.embedding_settings_widget.get_embedding_parameters()
-        selected_model = self.model_settings_widget.get_selected_model()
-        technique = embedding_params['technique']
-
-        QApplication.setOverrideCursor(Qt.WaitCursor)
-        progress_bar = ProgressBar(self, "Generating Embedding Visualization")
-        progress_bar.show()
-        
-        try:
-            # 2. Decide whether to use cached features or extract new ones
-            # This if/else block ONLY populates the 'features' variable.
-            if self.current_features is None or selected_model != self.current_feature_generating_model:
-                # SLOW PATH: Extract and cache new features
-                features, valid_data_items = self._extract_features(self.current_data_items, progress_bar=progress_bar)
-                
-                self.current_features = features
-                self.current_feature_generating_model = selected_model
-                self.current_data_items = valid_data_items
-                self.annotation_viewer.update_annotations(self.current_data_items)
-            else:
-                # FAST PATH: Use existing features
-                print("Using cached features. Skipping feature extraction.")
-                features = self.current_features
-
-            if features is None or len(features) == 0:
-                print("No valid features available. Aborting embedding.")
-                return
-
-            # 3. Run dimensionality reduction with the latest parameters
-            progress_bar.set_busy_mode(f"Running {technique} dimensionality reduction...")
-            embedded_features = self._run_dimensionality_reduction(features, embedding_params)
-            progress_bar.update_progress()
-            
-            if embedded_features is None:
-                return
-
-            # 4. Update the visualization with the new 2D layout
-            progress_bar.set_busy_mode("Updating visualization...")
-            self._update_data_items_with_embedding(self.current_data_items, embedded_features)
-            
-            self.embedding_viewer.update_embeddings(self.current_data_items)
-            self.embedding_viewer.show_embedding()
-            self.embedding_viewer.fit_view_to_points()
-            progress_bar.update_progress()
-            
-            print(f"Successfully generated embedding for {len(self.current_data_items)} annotations using {technique}")
-            
-        except Exception as e:
-            print(f"Error during embedding pipeline: {e}")
-            self.embedding_viewer.clear_points()
-            self.embedding_viewer.show_placeholder()
-            
-        finally:
-            QApplication.restoreOverrideCursor()
-            progress_bar.finish_progress()
-            progress_bar.stop_progress()
-            progress_bar.close()
-
-    def refresh_filters(self):
-        """Refresh display: filter data and update annotation viewer."""
-        QApplication.setOverrideCursor(Qt.WaitCursor)
-        try:
-            # Get filtered data and store for potential embedding
-            self.current_data_items = self.get_filtered_data_items()
-            
-            # --- MODIFIED: Invalidate the feature cache ---
-            # Since the filtered items have changed, the old features are no longer valid.
-            self.current_features = None
-            
-            # Update annotation viewer with filtered data
-            self.annotation_viewer.update_annotations(self.current_data_items)
-            
-            # Clear embedding viewer and show placeholder, as it is now out of sync
-            self.embedding_viewer.clear_points()
-            self.embedding_viewer.show_placeholder()
-            
-        finally:
-            QApplication.restoreOverrideCursor()
-
-    def on_label_selected_for_preview(self, label):
-        """Handle label selection to update preview state."""
-        if hasattr(self, 'annotation_viewer') and self.annotation_viewer.selected_widgets:
-            self.annotation_viewer.apply_preview_label_to_selected(label)
-            self.update_button_states()
-
-    def clear_preview_changes(self):
-        """Clear all preview changes and revert to original labels."""
-        if hasattr(self, 'annotation_viewer'):
-            self.annotation_viewer.clear_preview_states()
-            self.update_button_states()
-            print("Cleared all preview changes")
-
-    def update_button_states(self):
-        """Update the state of Clear Preview and Apply buttons."""
-        has_changes = (hasattr(self, 'annotation_viewer') and self.annotation_viewer.has_preview_changes())
-        
-        self.clear_preview_button.setEnabled(has_changes)
-        self.apply_button.setEnabled(has_changes)
-        
-        summary = self.annotation_viewer.get_preview_changes_summary()
-        self.clear_preview_button.setToolTip(f"Clear all preview changes - {summary}")
-        self.apply_button.setToolTip(f"Apply changes - {summary}")
-
-    def apply(self):
-        """Apply any modifications to the actual annotations."""
-        # Make cursor busy
-        QApplication.setOverrideCursor(Qt.WaitCursor)
-        
-        try:
-            applied_annotations, annotations_to_delete = self.annotation_viewer.apply_preview_changes_permanently()
-            
-            if applied_annotations:
-                # Find which data items were affected and tell their visual components to update
-                changed_ids = {ann.id for ann in applied_annotations}
-                for item in self.current_data_items:
-                    if item.annotation.id in changed_ids:
-                        # Update annotation widget in the grid
-                        widget = self.annotation_viewer.annotation_widgets_by_id.get(item.annotation.id)
-                        if widget:
-                            widget.update() # Repaint to show new permanent color
-
-                        # Update point in the embedding viewer
-                        point = self.embedding_viewer.points_by_id.get(item.annotation.id)
-                        if point:
-                            point.update() # Repaint to show new permanent color
-
-                # Update sorting if currently sorting by label
-                if self.annotation_viewer.sort_combo.currentText() == "Label":
-                    self.annotation_viewer.recalculate_widget_positions()
-
-                # Update the main application's data
-                affected_images = {ann.image_path for ann in applied_annotations}
-                for image_path in affected_images:
-                    self.image_window.update_image_annotations(image_path)
-                self.annotation_window.load_annotations()
-                
-                print(f"Applied label changes to {len(applied_annotations)} annotation(s)")
-
-            if annotations_to_delete:
-                # Actually delete the annotations using AnnotationWindow's method
-                self.annotation_window.delete_annotations(annotations_to_delete)
-                
-                # Remove deleted annotations from our data structures
-                deleted_ids = {ann.id for ann in annotations_to_delete}
-                
-                # Remove from current_data_items
-                self.current_data_items = [item for item in self.current_data_items 
-                                         if item.annotation.id not in deleted_ids]
-                
-                # Remove widgets from annotation viewer
-                for ann_id in deleted_ids:
-                    widget = self.annotation_viewer.annotation_widgets_by_id.pop(ann_id, None)
-                    if widget:
-                        widget.setParent(None)
-                        widget.deleteLater()
-                
-                # Remove points from embedding viewer
-                for ann_id in deleted_ids:
-                    point = self.embedding_viewer.points_by_id.pop(ann_id, None)
-                    if point:
-                        self.embedding_viewer.graphics_scene.removeItem(point)
-                
-                # Update layouts
-                self.annotation_viewer.recalculate_widget_positions()
-                
-                print(f"Deleted {len(annotations_to_delete)} annotation(s)")
-                
-            # Clear selection and button states
-            self.annotation_viewer.clear_selection()
-            self.embedding_viewer.render_selection_from_ids(set()) # Clear embedding selection
-            self.update_button_states()
-            
-            if applied_annotations or annotations_to_delete:
-                total_changes = len(applied_annotations) + len(annotations_to_delete)
-                print(f"Applied {total_changes} total change(s)")
-            else:
-                print("No preview changes to apply")
-
-        except Exception as e:
-            print(f"Error applying modifications: {e}")
-        finally:
-            # Restore cursor
-            QApplication.restoreOverrideCursor()
-            
     def _cleanup_resources(self):
-        """
-        Clean up heavy resources like the loaded model and clear GPU cache.
-        This is called when the window is closed to free up memory.
-        """
+        """Frees up heavy resources when the window is closed."""
         print("Cleaning up Explorer resources...")
-        
-        # Reset model and feature caches
-        self.imgsz = 128
+        self.embedding_viewer.animation_timer.stop()
         self.loaded_model = None
         self.model_path = ""
         self.current_features = None
         self.current_feature_generating_model = ""
-        
-        # Clear CUDA cache if available to free up GPU memory
+
         if torch.cuda.is_available():
             print("Clearing CUDA cache.")
             torch.cuda.empty_cache()
-            
+
         print("Cleanup complete.")
